Generic Analyses

The Analyses class is used to harmonize the interface to data or links to raw data files and results across different types of data. Analyses child classes are used for specific data types, providing dedicated methods.

Usage

Analyses(analyses = list(), results = list())

# S3 method for class 'Analyses'
validate_object(x)

Arguments

analyses

A list of analyses, where each element is a data entry or a connection to a raw data file.

results

A list of results, where each element is a specific Results child class.

x

An Analyses object.

Value

An object of class Analyses, which is a list containing:

analyses: A list of analyses, where each element is a data entry or a connection to a raw data file.

results: A list of results, where each element is a specific Results child class.

type: A character string indicating the type of data contained in the analyses.

formats: A character vector indicating the possible formats of the analyses.

Methods (by generic)

validate_object(Analyses): Validate the Analyses object, returning NULL if valid or an error if not.