Generic Entry for Audit Trail

The AuditTrailEntry class is used as an element in the AuditTrail entries.

Usage

AuditTrailEntry(time_stamp, value_class, value_parent, value)

Arguments

time_stamp

A POSIXct timestamp of the entry.

value_class

A character string representing the class of the entry.

value_parent

A character string representing the parent class of the entry.

value

A character string representing the value of the entry.