Configuration Parameter for Caching

The ConfigCache class is a ConfigParameter for the caching behavior.

Usage

ConfigCache(
  value = TRUE,
  mode = "rds",
  folder = "cache",
  file = "cache.sqlite"
)

# S3 method for class 'ConfigCache'
validate_object(x)

# S3 method for class 'ConfigCache'
size(x)

# S3 method for class 'ConfigCache'
info(x)

Arguments

value

Logical indicating whether to enable or disable caching.

mode

Character indicating the caching mode (e.g., "rds" or "sqlite").

folder

Character indicating the folder for caching (for "rds" mode).

file

Character indicating the file for caching (for "sqlite" mode).

x

A ConfigCache object.

Value

A ConfigCache object, which is a list with the following elements:

name: Name of the cache configuration.

description: Description of the cache configuration.

value: Logical value indicating whether caching is enabled.

mode: Character string indicating the caching mode (e.g., "rds" or "sqlite").

folder: Character string indicating the folder for caching (for "rds" mode).

file: Character string indicating the file for caching (for "sqlite" mode).

Methods (by generic)

validate_object(ConfigCache): Validates a ConfigCache object, returning NULL if valid.

size(ConfigCache): Get the size of the cache as a named numeric vector with the size of the cache in bytes, KB, MB, or GB.

info(ConfigCache): Get information about the cache as a data.table with the names and number of rows in each table for SQLite cache, or a list of files for RDS cache.