Engine Metadata

The EngineMetadata class is a subclass of the Metadata class. It is used to store metadata of data processing engines.

Usage

EngineMetadata(entries = list(), engine = "CoreEngine")

Arguments

entries

Named list with metadata entries. Note that all given elements must be named and of length one. If an element "name" is given, it must be type character. If an element "date" is given, it must be class POSIXct or POSIXt. If given "date" is character, conversion to class POSIXct or POSIXt is attempted.

engine

A character string with the name of the engine. Default is "CoreEngine".

Slots

entries

List object with metadata entries as elements of length one.

engine

A character string with the name of the engine.

See also