Mass Spectrometry Analyses

The MassSpecAnalyses class represents mass spectrometry (MS) raw data files and holds results from processing MS data. It is a subclass of the Analyses class and provides methods to manage and inspect MS data. The MassSpecAnalyses class is built from a character vector of file paths to MS raw data files. The possible file formats are mzML and mzXML. If msconvert from ProteoWizard is installed and found via CLI (i.e., must be added to the environmental variables), the engine can also load vendor formats (i.e., Agilent MassHunter .d, Thermo Scientific RAW, Shimadzu LCD (except ITOF), Sciex WIFF/WIFF2) by direct conversion to mzML. Note that conversion of vendor formats is only possible under Windows OS.

Usage

MassSpecAnalyses(files = NULL, centroid = FALSE, levels = c(1, 2))

Arguments

files

A character vector with the file paths of MS raw data files.

centroid

Logical (length 1). Set to TRUE to centroid data when converting from vendor formats to mzML.

levels

Numeric vector with the MS levels to consider when centroiding data. Default is c(1, 2).

Slots

analyses

(getter/setter) A list of objects representing individual mass spectrometry raw data files.

replicates

(getter/setter) A character vector representing the names of the replicate name for each analysis.

blanks

(getter/setter) A character vector representing the names of the blank name for each analysis.

concentrations

(getter/setter) A numeric vector representing the concentration for each analysis.

types

(getter) A character vector representing the type of each analysis.

files

(getter) A character vector with the file path of each analysis.

formats

(getter) A character vector with the file format of each analysis.

instruments

(getter) A list with the instrument information for each analysis.

software

(getter) A list with the software information for each analysis.

spectra_number

(getter) A numeric vector with the number of spectra for each analysis.

spectra_headers

(getter) A list with the headers of the spectra for each analysis.

spectra_mode

(getter) A character vector with the mode of the spectra for each analysis.

spectra_level

(getter) A character vector with the level of the spectra for each analysis.

spectra_lowest_mz

(getter) A numeric vector with the lowest m/z value for each analysis.

spectra_highest_mz

(getter) A numeric vector with the highest m/z value for each analysis.

spectra_lowest_rt

(getter) A numeric vector with the lowest retention time for each analysis.

spectra_highest_rt

(getter) A numeric vector with the highest retention time for each analysis.

spectra_lowest_mobility

(getter) A numeric vector with the lowest mobility for each analysis.

spectra_highest_mobility

(getter) A numeric vector with the highest mobility for each analysis.

spectra_polarity

(getter) A character vector with the polarity of the spectra for each analysis.

spectra_tic

(getter) A list with the total ion current (TIC) for each analysis.

spectra_bpc

(getter) A list with the base peak chromatogram (BPC) for each analysis.

chromatograms_number

(getter) A numeric vector with the number of chromatograms for each analysis.

chromatograms_headers

(getter) A list with the headers of the chromatograms for each analysis.

chromatograms_raw

(getter) A list with the raw chromatograms for each analysis.

info

(getter) A data frame with the information of each analysis.

has_ion_mobility

(getter) A logical vector indicating if the analyses have ion mobility data.

has_results_nts

(getter) A logical vector indicating if the analyses have NTS results.

has_results_spectra

(getter) A logical vector indicating if the analyses have spectra results.

has_results_chromatograms

(getter) A logical vector indicating if the analyses have chromatograms results.

NTS

(getter/setter) A list with the NTS results for each analysis.

Spectra

(getter/setter) A list with the spectra results for each analysis.

Chromatograms

(getter/setter) A list with the chromatograms results for each analysis.

References

Kapoulkine A (2022). “pugixml 1.13: Light-weight, simple and fast XML parser for C++ with XPath support.” Copyright (C) 2006-2018. http://pugixml.org.

Kessner D, Chambers M, Burke R, Agus D, Mallick P (2008). “ProteoWizard: open source software for rapid proteomics tools development.” Bioinformatics, 24(21), 2534–2536. doi:10.1093/bioinformatics/btn323 , http://www.ncbi.nlm.nih.gov/pubmed/18606607.

Chambers MC, Maclean B, Burke R, Amodei D, Ruderman DL, Neumann S, Gatto L, Fischer B, Pratt B, Egertson J, Hoff K, Kessner D, Tasman N, Shulman N, Frewen B, Baker TA, Brusniak M, Paulse C, Creasy D, Flashner L, Kani K, Moulding C, Seymour SL, Nuwaysir LM, Lefebvre B, Kuhlmann F, Roark J, Rainer P, Detlev S, Hemenway T, Huhmer A, Langridge J, Connolly B, Chadick T, Holly K, Eckels J, Deutsch EW, Moritz RL, Katz JE, Agus DB, MacCoss M, Tabb DL, Mallick P (2012). “A cross-platform toolkit for mass spectrometry and proteomics.” Nature Biotechnology, 30(10), 918–920. https://doi.org/10.1038/nbt.2377.