MassSpecAnalysis S3 class constructor, methods and functions

Creates a MassSpecAnalysis S3 class object.

Usage

MassSpecAnalysis(
  name = NA_character_,
  replicate = NA_character_,
  blank = NA_character_,
  file = NA_character_,
  format = NA_character_,
  type = NA_character_,
  time_stamp = NA_character_,
  spectra_number = NA_integer_,
  spectra_mode = NA_character_,
  spectra_levels = NA_integer_,
  mz_low = NA_real_,
  mz_high = NA_real_,
  rt_start = NA_real_,
  rt_end = NA_real_,
  polarity = NA_character_,
  has_ion_mobility = FALSE,
  chromatograms_number = NA_integer_,
  software = data.table(),
  instrument = data.table(),
  run = data.table(),
  spectra = data.table(),
  chromatograms = data.table(),
  features_eic = list(),
  features = data.table(),
  metadata = list(),
  version = NA_character_
)

# S3 method for MassSpecAnalysis
validate(x = NULL)

# S3 method for MassSpecAnalysis
print(x, ...)

# S3 method for MassSpecAnalysis
asJSON(x)

as.MassSpecAnalysis(value)

parse.MassSpecAnalysis(files = NULL, runParallel = FALSE)

Arguments

name

mzML or mzXML file name without extension.

replicate

Character with length one, representing the analysis replicate group name.

blank

Character with length one, representing the associated blank replicate group name.

file

mzML or mzXML full file path (with extension).

format

Character with length one. One of mzML or mzXML.

type

Character with length one. Possible values as "MS" for only MS1 spectral data, "MS/MS" for tandem spectral data (i.e., MS1 and MS2) or "SRM" for selected reaction monitoring data (i.e., no spectra only chromatograms).

time_stamp

Character with length one, representing the start time and date of the data acquisition.

spectra_number

Integer with the number of spectra in the file.

spectra_mode

Character with length one. Possible values are centroid , profile or NA for centroided data, data in profile mode or mode not defined or absence spectra, respectively.

spectra_levels

Integer vector with the number of levels in the spectra (i.e., 1 or 1 and 2).

mz_low

Numeric length one with the lowest m/z value detected in the spectra, in Da.

mz_high

Numeric length one with the highest m/z value detected in the spectra, in Da.

rt_start

Numeric length one with the run start time, in seconds.

rt_end

Numeric length one with the run end time, in seconds.

polarity

Character length one. Possible values are positive, negative or both (the latter refers to polarity switching acquisition).

has_ion_mobility

Logical length one for presence or absence of drift separation from ion mobility.

chromatograms_number

Integer with the number of chromatograms in the file.

software

data.table with software info used to acquire/process the data (content is highly vendor dependent).

instrument

data.table with instrument info used to acquire the data (content is highly vendor dependent).

run

data.table run information for each spectrum.

spectra

data.table with the raw spectra (only present if loaded).

chromatograms

data.table with the raw chromatograms (only present if loaded).

features_eic

list with the feature extracted ions chromatograms (EICs) from binning of spectra.

features

data.table with the features from data processing.

metadata

List with flexible storage for experimental metadata (e.g., concentration, location, etc.).

version

Character of length one with the version. It should match with the version of the StreamFind package when created the MassSpecAnalysis object.

x

A MassSpecAnalysis S3 class object.

...

Not used.

value

A list to be checked and/or converted to MassSpecAnalysis S3 class.

files

A character vector with mzML or mzXML full file path/s. Alternatively, a data.frame with the column/s file, replicate and blank with the full file path/s, the replicate group name/s (string) and the associated blank replicate group name (string).

runParallel

Logical of length one. Set to TRUE for parallel processing. Note that parallel processing is not always faster due to overhead.

Value

An MassSpecAnalysis S3 class object.

Methods (by generic)

validate(MassSpecAnalysis): Validates a MassSpecAnalysis S3 class object, returning a logical value of length one.

print(MassSpecAnalysis): Prints the MassSpecAnalysis S3 class object in the console.

asJSON(MassSpecAnalysis): Converts a MassSpecAnalysis S3 class object to a JSON string.

Functions

as.MassSpecAnalysis(): Converts the argument value in a MassSpecAnalysis S3 class object.

parse.MassSpecAnalysis(): Parses information from mzML or mzXML file/s and returns a list with MassSpecAnalysis S3 class object/s. On error, returns NULL.