MassSpecMethod_BinSpectra_StreamFind Class

Bins spectral data based on variables.

Usage

MassSpecMethod_BinSpectra_StreamFind(
  binNames = c("rt", "mz"),
  binValues = c(5, 10),
  byUnit = TRUE,
  refBinAnalysis = NA_integer_
)

Arguments

binNames

Character with the variable names for building the bins. Possible values are rt, mz, mass, mobility. Note that the binNames must be in the spectra column names.

binValues

Numeric with the bin values for each variable.

byUnit

Logical of length one to bin by unit, meaning that the binning is performed by the number of units not actual values. For instance, if the bin is 10 and binName is rt, then the binning is performed by 10 seconds not 10 values for each bin. If byUnit is FALSE, then the binning is performed by the actual values but only the first of binNames and binValues is used.

refBinAnalysis

The analysis index to use a reference for creating the bins.

Value

A MassSpecMethod_BinSpectra_StreamFind object.