Mass Spectrometry Method to Calculate Features Quality (StreamFind algorithm)

Settings for calculating quality parameters of features (e.g., signal-to-noise (sn) ratio).

Usage

MassSpecMethod_CalculateFeaturesQuality_StreamFind(
  filtered = FALSE,
  rtExpand = 0,
  mzExpand = 0,
  minPeakWidth = 6,
  maxPeakWidth = 30,
  minTracesIntensity = 0,
  minNumberTraces = 6,
  baseCut = 0
)

Arguments

filtered

Logical (length 1). When TRUE, filtered features/groups are considered.

rtExpand

Numeric (length 1) with the retention time values (in seconds) for expanding left and right the retention time range.

mzExpand

Numeric (length 1) with the mass or mass-to-charge ratio (m/z) (in Da) for expanding left and right the mass range.

minPeakWidth

Numeric of length one with the minimum peak width centered on the maximum for extracting the feature EIC.

maxPeakWidth

Numeric of length one with the maximum peak width centered on the maximum for reconstructing the feature EIC.

minTracesIntensity

Numeric of length 1 with the minimum intensity of spectra traces for calculating feature quality.

minNumberTraces

Numeric of length 1 with the minimum number traces for calculating feature quality.

baseCut

Numeric of length 1 with the base cut for calculating feature Gaussian fit.

Value

A MassSpecMethod_CalculateFeaturesQuality_StreamFind object.