MassSpecMethod_CorrectMatrixSuppression_TiChri Class

Settings for correcting matrix suppression based on the TiChri algorithm from Tisler et al. (2021). The algorithm calculates the matrix profile for the total ion chromatogram (TIC) and corrects the matrix suppression for features. Internal standards can be assigned to improve the correction. The suppression_factor is added to the feature list and can be used to correct the features intensity. The argument/parameter correctIntensity is available in plotting and processing methods and when TRUE, the suppression factor is used to correct the feature intensity for better comparison across analyses with different matrix suppression.

Usage

MassSpecMethod_CorrectMatrixSuppression_TiChri(
  mpRtWindow = 10,
  istdAssignment = "none",
  istdRtWindow = 5,
  istdN = 2
)

Arguments

mpRtWindow

Numeric of length one with the retention time window (in seconds) for calculating the matrix profile.

istdAssignment

Character of length one with the assignment method for internal standards. Possible values are "nearest", "range", and "none". Default is "nearest". Setting "nearest" assigns the nearest istdN internal standard/s, "range" assigns internal standard/s within the istdRtWindow window, and "none" does not assign internal standards. If internal standards are assigned, the tichri value is calculated for each internal standard and used to correct the matrix suppression for the features. If no internal standards are assigned, the correction is based only on the TIC matrix profile, which is less accurate.

istdRtWindow

Numeric of length one with the retention time window (in seconds) for assigning internal standards. Default is 5.

istdN

Integer of length one with the number of internal standards to assign. Default is 2.

Value

A MassSpecMethod_CorrectMatrixSuppression_TiChri object.

References

Tisler S, Pattison DI, Christensen JH (2021). “Correction of Matrix Effects for Reliable Non-target Screening LC–ESI–MS Analysis of Wastewater.” Analytical Chemistry, 93(24), 8432-8441. doi:10.1021/acs.analchem.1c00357 , PMID: 34096716, https://doi.org/10.1021/acs.analchem.1c00357.