MassSpecMethod_FillFeatures_StreamFind S7 class

Settings for filling missing values in features.

Usage

MassSpecMethod_FillFeatures_StreamFind(
  withinReplicate = TRUE,
  filtered = TRUE,
  rtExpand = 0,
  mzExpand = 0,
  minPeakWidth = 6,
  maxPeakWidth = 30,
  minTracesIntensity = 1000,
  minNumberTraces = 5,
  minIntensity = 5000,
  baseCut = 0.3,
  maxSearchWindow = 5,
  minSignalToNoiseRatio = 3,
  minGaussianFit = 0.2
)

Arguments

withinReplicate

Logical of length one to fill within replicates not global.

filtered

Logical of length one to consider filtered features or not.

rtExpand

Numeric (length 1) with the retention time values (in seconds) for expanding left and right the retention time range.

mzExpand

Numeric (length 1) with the mass or mass-to-charge ratio (m/z) (in Da) for expanding left and right the mass range.

minPeakWidth

Numeric of length one with the minimum peak width for building feature extraction targets.

maxPeakWidth

Numeric of length one with the maximum peak width for building feature extraction targets.

minTracesIntensity

Numeric of length one with the minimum intensity to collect spectra data for extracted ion chromatograms.

minNumberTraces

Integer of length one with the minimum number of traces to consider a feature.

minIntensity

Numeric of length one with the minimum intensity to consider a feature. Meaning that the maximum intensity of the extracted ion chromatogram must be greater than the defined value.

baseCut

Numeric of length one with the base cut for building Gaussian model.

maxSearchWindow

Numeric of length one with maximum time search window (seconds) from the averaged group retention time. Default to 5 seconds.

minSignalToNoiseRatio

Numeric of length one with the minimum signal to noise ratio to consider a feature.

minGaussianFit

Numeric of length one with the minimum Gaussian fit to consider a feature.

Value

A MassSpecMethod_FillFeatures_StreamFind class object.