MassSpecMethod_FilterFeaturesMS2_native S7 class

Settings for filtering (i.e., cleaning) non-relevant traces in features MS2 spectrum.

Usage

MassSpecMethod_FilterFeaturesMS2_native(
  top = NULL,
  minIntensity = NULL,
  relMinIntensity = NULL,
  blankClean = FALSE,
  mzClust = 0.005,
  blankPresenceThreshold = 0.8,
  globalPresenceThreshold = 0.1
)

Arguments

top

An integer specifying the number of top features to be kept.

minIntensity

A numeric value specifying the minimum intensity.

relMinIntensity

A numeric value specifying the relative minimum intensity.

blankClean

A logical value specifying if blank cleaning should be performed. When TRUE MS2 traces present in the blank samples above the defined blankPresenceThreshold and above the globalPresenceThreshold are excluded.

mzClust

A numeric value specifying the m/z clustering threshold.

blankPresenceThreshold

A numeric value specifying the blank presence threshold.

globalPresenceThreshold

A numeric value specifying the global presence threshold.

Value

A MassSpecMethod_FilterFeaturesMS2_native object.