MassSpecMethod_FilterFeatures_StreamFind S7 class

Settings for filtering of features and feature groups.

Usage

MassSpecMethod_FilterFeatures_StreamFind(
  correctIntensity = TRUE,
  minSnRatio = NA_real_,
  minGaussianFit = NA_real_,
  excludeIsotopes = FALSE,
  excludeAdducts = FALSE,
  minIntensity = NA_real_,
  maxDeviationInReplicate = NA_real_,
  minAbundanceInReplicate = NA_real_,
  blankThreshold = NA_real_,
  conservative = TRUE,
  onlyWithMS2 = FALSE
)

Arguments

correctIntensity

Logical (length 1) with TRUE for correcting the intensity for matrix suppression before applying intensity based filters. Note that suppression factor as obtained by the processing method CorrectMatrixSuppression must be present. When not available the correction is ignored even if set to TRUE.

minSnRatio

Numeric (length 1) with the minimum signal-to-noise ratio.

minGaussianFit

Numeric (length 1) with the minimum Gaussian fit.

excludeIsotopes

Logical (length 1) with TRUE for filtering annotated isotopes (only prevails the monoisotopic features).

excludeAdducts

Logical (length 1) with TRUE for filtering annotated adducts.

minIntensity

Numeric (length 1) with the minimum intensity threshold.

maxDeviationInReplicate

Numeric (length 1) with the maximum standard deviation of the intensity in each replicate group. The value must be a percentage from 0 to 100. The filter is applied if the minimum standard deviation from all replicates is above the value.

minAbundanceInReplicate

Numeric (length 1) with the minimum abundance in each replicate group. The filter is applied if the maximum abundance from all replicates is below the value.

blankThreshold

Numeric (length 1) with the intensity multiplier to set the blank threshold. The filter is applied if the maximum mean intensity from all non-blank replicates is below the correspondent blank intensity times the multiplier.

conservative

Logical (length 1) with TRUE for conservative filtering. The filters are applied on a feature group basis. If FALSE, the filters are applied on a feature basis. This means that when TRUE a feature is only filtered if all features within a feature group comply with the filter.

onlyWithMS2

Logical (length 1) with TRUE for filtering features without MS2 spectra.

Value

A MassSpecMethod_FilterFeatures_StreamFind object.