MassSpecMethod_FindFeatures_openms class

Settings for finding features (i.e., chromatographic peaks) in mzML/mzXML files using the OpenMS (https://abibuilder.cs.uni-tuebingen.de/archive/openms/) software with the algorithm FeatureFinderMetabo. The function uses the package patRoon in the background.

Usage

MassSpecMethod_FindFeatures_openms(
  noiseThrInt = 1000,
  chromSNR = 3,
  chromFWHM = 7,
  mzPPM = 15,
  reEstimateMTSD = TRUE,
  traceTermCriterion = "sample_rate",
  traceTermOutliers = 5,
  minSampleRate = 1,
  minTraceLength = 4,
  maxTraceLength = 70,
  widthFiltering = "fixed",
  minFWHM = 4,
  maxFWHM = 35,
  traceSNRFiltering = TRUE,
  localRTRange = 0,
  localMZRange = 0,
  isotopeFilteringModel = "none",
  MZScoring13C = FALSE,
  useSmoothedInts = FALSE,
  intSearchRTWindow = 3,
  useFFMIntensities = FALSE,
  verbose = FALSE
)

# S3 method for class 'MassSpecMethod_FindFeatures_openms'
validate_object(x)

Arguments

noiseThrInt

Intensity threshold below which peaks are regarded as noise.

chromSNR

Minimum signal-to-noise a mass trace should have.

chromFWHM

Expected chromatographic peak width (in seconds).

mzPPM

Allowed mass deviation (in ppm).

reEstimateMTSD

Enables dynamic re-estimation of m/z variance during mass trace collection stage.

traceTermCriterion

Termination criterion for the extension of mass traces. In 'outlier' mode, trace extension cancels if a predefined number of consecutive outliers are found (see trace_termination_outliers parameter). In 'sample_rate' mode, trace extension in both directions stops if ratio of found peaks versus visited spectra falls below the 'min_sample_rate' threshold.

traceTermOutliers

Mass trace extension in one direction cancels if this number of consecutive spectra with no detectable peaks is reached.

minSampleRate

Minimum fraction of scans along the mass trace that must contain a peak.

minTraceLength

Minimum expected length of a mass trace (in seconds).

maxTraceLength

Maximum expected length of a mass trace (in seconds). Set to a negative value to disable maximal length check during mass trace detection.

widthFiltering

Enable filtering of unlikely peak widths. The fixed setting filters out mass traces outside the min_fwhm, max_fwhm interval (set parameters accordingly!). The auto setting filters with the 5 and 95% quantiles of the peak width distribution.

minFWHM

Minimum full-width-at-half-maximum of chromatographic peaks (in seconds). Ignored if parameter width_filtering is off or auto.

maxFWHM

Maximum full-width-at-half-maximum of chromatographic peaks (in seconds). Ignored if parameter width_filtering is off or auto.

traceSNRFiltering

Apply post-filtering by signal-to-noise ratio after smoothing.

localRTRange

RT range where to look for coeluting mass traces.

localMZRange

MZ range where to look for isotopic mass traces.

isotopeFilteringModel

Remove/score candidate assemblies based on isotope intensities. SVM isotope models for metabolites were trained with either 2% or 5% RMS error. For peptides, an averagine cosine scoring is used. Select the appropriate noise model according to the quality of measurement or MS device.

MZScoring13C

Use the 13C isotope peak position (~1.003355 Da) as the expected shift in m/z for isotope mass traces (highly recommended for lipidomics!). Disable for general metabolites (as described in Kenar et al. 2014, MCP.).

useSmoothedInts

Use LOWESS intensities instead of raw intensities.

intSearchRTWindow

Retention time window (in seconds, +/- feature retention time) that is used to find the closest data point to the retention time to obtain the intensity of a feature (this is needed since OpenMS does not provide this data).

useFFMIntensities

If TRUE then peak intensities are directly loaded from FeatureFinderMetabo output. Otherwise, intensities are loaded afterwards from the input ‘mzML’ files, which is potentially much slower, especially with many analyses files. However, useFFMIntensities=TRUE is still somewhat experimental, may be less accurate and requires a recent version of OpenMS (>=2.7).

verbose

Logical of length one. When TRUE adds processing information to the console.

x

A MassSpecMethod_FindFeatures_openms object.

Value

A MassSpecMethod_FindFeatures_openms object.

Details

See the findFeaturesOpenMS function from the patRoon package for more information and requirements.

Methods (by generic)

validate_object(MassSpecMethod_FindFeatures_openms): Validate the MassSpecMethod_FindFeatures_openms object, returning NULL if valid.

References

Helmus R, ter Laak TL, van Wezel AP, de Voogt P, Schymanski EL (2021). “patRoon: open source software platform for environmental mass spectrometry based non-target screening.” Journal of Cheminformatics, 13(1). doi:10.1186/s13321-020-00477-w .

Helmus R, van de Velde B, Brunner AM, ter Laak TL, van Wezel AP, Schymanski EL (2022). “patRoon 2.0: Improved non-target analysis workflows including automated transformation product screening.” Journal of Open Source Software, 7(71), 4029. doi:10.21105/joss.04029 .

Röst HL, Sachsenberg T, Aiche S, Bielow C, Weisser H, Aicheler F, Andreotti S, Ehrlich H, Gutenbrunner P, Kenar E, Liang X, Nahnsen S, Nilse L, Pfeuffer J, Rosenberger G, Rurik M, Schmitt U, Veit J, Walzer M, Wojnar D, Wolski WE, Schilling O, Choudhary JS, Malmström L, Aebersold R, Reinert K, Kohlbacher O (2016). “OpenMS: a flexible open-source software platform for mass spectrometry data analysis.” Nature Methods, 13(9), 741–748. doi:10.1038/nmeth.3959 .