MassSpecMethod_FindFeatures_qalgorithms S7 class

The qAlgorithms uses a comprehensive peak model developed by Renner et al. to identify peaks within LC-MS data. More information can be found in the GitHub repository of the qAlgorithms project. The qAlgorithms is best used with profile data but centroid data is also possible. Yet, a mass uncertainty should by supplied in parts per million (ppm) to account for the m/z deviation in centroid data.

Usage

MassSpecMethod_FindFeatures_qalgorithms(ppm = 5)

Arguments

ppm

numeric(1) defining the maximal tolerated m/z deviation in parts per million (ppm) only applicable for centroid data. For profile data, the ppm value is ignored.