MassSpecMethod_GroupFeatures_xcms3_peakdensity Class

Settings for grouping features (i.e., chromatographic peaks) across mzML/mzXML files using the package xcms (version 3) with the algorithm peakDensity. The function uses the package patRoon in the background.

Usage

MassSpecMethod_GroupFeatures_xcms3_peakdensity(
  bw = 5,
  minFraction = 1,
  minSamples = 1,
  binSize = 0.008,
  maxFeatures = 100
)

Arguments

bw

numeric(1) defining the bandwidth (standard deviation of the smoothing kernel) to be used. This argument is passed to the density() method.

minFraction

numeric(1) defining the minimum fraction of analyses in at least one analysis replicate group in which the features have to be present to be considered as a feature group.

minSamples

numeric(1) with the minimum number of analyses in at least one analysis replicate group in which the features have to be detected to be considered a feature group.

binSize

numeric(1) defining the size of the overlapping slices in mz dimension.

maxFeatures

numeric(1) with the maximum number of feature groups to be identified in a single mz slice.

Value

A MassSpecMethod_GroupFeatures_xcms3_peakdensity object.

Details

See the groupFeaturesXCMS3 function from the patRoon package for more information and requirements.

References

Helmus R, ter Laak TL, van Wezel AP, de Voogt P, Schymanski EL (2021). “patRoon: open source software platform for environmental mass spectrometry based non-target screening.” Journal of Cheminformatics, 13(1). doi:10.1186/s13321-020-00477-w .

Helmus R, van de Velde B, Brunner AM, ter Laak TL, van Wezel AP, Schymanski EL (2022). “patRoon 2.0: Improved non-target analysis workflows including automated transformation product screening.” Journal of Open Source Software, 7(71), 4029. doi:10.21105/joss.04029 .

Smith, C.A., Want, E.J., O'Maille, G., Abagyan,R., Siuzdak, G. (2006). “XCMS: Processing mass spectrometry data for metabolite profiling using nonlinear peak alignment, matching and identification.” Analytical Chemistry, 78, 779–787.

Tautenhahn R, Boettcher C, Neumann S (2008). “Highly sensitive feature detection for high resolution LC/MS.” BMC Bioinformatics, 9, 504.

Benton HP, Want EJ, Ebbels TMD (2010). “Correction of mass calibration gaps in liquid chromatography-mass spectrometry metabolomics data.” BIOINFORMATICS, 26, 2488.