MassSpecMethod_IntegrateChromatograms_pracma Class

Integrates chromatograms using the function findpeaks from the package pracma with natively added peak exclusion and evaluation steps.

Usage

MassSpecMethod_IntegrateChromatograms_pracma(
  merge = TRUE,
  closeByThreshold = 45,
  minPeakHeight = 0,
  minPeakDistance = 10,
  minPeakWidth = 5,
  maxPeakWidth = 120,
  minSN = 10
)

# S3 method for class 'MassSpecMethod_IntegrateChromatograms_pracma'
validate_object(x)

Arguments

merge

Logical (length 1) indicating if the nearby peaks should be merged.

closeByThreshold

Numeric (length 1) with the maximum distance between peaks to be merged.

minPeakHeight

Numeric (length 1) with the minimum peak height to be considered.

minPeakDistance

Numeric (length 1) with the minimum distance between peaks.

minPeakWidth

Numeric (length 1) with the minimum peak width.

maxPeakWidth

Numeric (length 1) with the maximum peak width.

minSN

Numeric (length 1) with the minimum signal-to-noise ratio.

x

A MassSpecMethod_IntegrateChromatograms_pracma object.

Value

A MassSpecMethod_IntegrateChromatograms_pracma object.

Methods (by generic)

validate_object(MassSpecMethod_IntegrateChromatograms_pracma): Validate the MassSpecMethod_IntegrateChromatograms_pracma object, returning NULL if valid.