MassSpecMethod_LoadSpectra_native

.

Usage

MassSpecMethod_LoadSpectra_native(
  levels = 1,
  mzmin = 0,
  mzmax = 0,
  rtmin = 0,
  rtmax = 0,
  mobilitymin = 0,
  mobilitymax = 0,
  minIntensity = 0
)

Arguments

levels

A numeric vector with the levels to be used.

mzmin

A numeric vector with the minimum m/z values to be used.

mzmax

A numeric vector with the maximum m/z values to be used.

rtmin

A numeric vector with the minimum retention time values to be used.

rtmax

A numeric vector with the maximum retention time values to be used.

mobilitymin

A numeric vector with the minimum mobility values to be used.

mobilitymax

A numeric vector with the maximum mobility values to be used.

minIntensity

A numeric value with the minimum intensity to be used.

Value

A MassSpecMethod_LoadSpectra_native object.