MassSpecMethod_SuspectScreening_StreamFind Class

Processing method for suspect screening using a data.frame with target compounds.

Usage

MassSpecMethod_SuspectScreening_StreamFind(
  database = data.table::data.table(name = character(), formula = character(), mass =
    numeric()),
  ppm = 5,
  sec = 10,
  ppmMS2 = 10,
  mzrMS2 = 0.008,
  minCusiness = 0.5,
  minFragments = 3,
  filtered = FALSE
)

# S3 method for class 'MassSpecMethod_SuspectScreening_StreamFind'
validate_object(x)

Arguments

database

A data.frame with at least the columns name and mass, indicating the name and neutral monoisotopic mass of the suspect targets.

ppm

Numeric of length one with the mass deviation, in ppm.

sec

Numeric of length one with the time deviation, in seconds.

ppmMS2

Numeric (length 1) with the mass deviation for MS2, in ppm.

mzrMS2

Numeric length 1 with the absolute m/z range window for matching MS2 fragments with experimental MS2 traces.

minCusiness

Numeric length 1 with the minimum cusiness value of the experimental and database fragments for a feature to be considered a suspect.

minFragments

Integer (length 1) with the minimum number of fragments.

filtered

Logical (length 1). When TRUE, filtered features/groups are considered.

x

A MassSpecMethod_SuspectScreening_StreamFind object.

Value

A MassSpecMethod_SuspectScreening_StreamFind object.

Methods (by generic)

validate_object(MassSpecMethod_SuspectScreening_StreamFind): Validate the MassSpecMethod_SuspectScreening_StreamFind object, returning NULL if valid.