MassSpecMethod_SuspectScreening_forident

Settings for performing suspect screening using the FOR-IDENT platform.

Usage

MassSpecMethod_SuspectScreening_forident(
  addMS2 = FALSE,
  useNeutralMass = TRUE,
  path = getwd(),
  name = "feature_list"
)

Arguments

addMS2

Logical length 1. When TRUE and MS2 data is available, the fragments pattern (i.e., MS2 averaged spectra) is added to the .txt file to import in FOR-IDENT platform. Note that when addMS2 is TRUE the m/z values are used instead of neutral mass even is useNeutralMass is set to TRUE.

useNeutralMass

Logical length 1. When TRUE and neutral mass is available, the neutral mass of features/feature groups is used instead of the m/z.

path

Character length 1 with the path to save the .txt file with the list of features for identification.

name

Character length 1 with the name of the file (without extension) to be saved in the path.

Value

A MassSpecMethod_SuspectScreening_forident object.

Note

After processing, a .txt file as defined by name and path is created with the list of features or feature groups to be imported in the FOR-IDENT platform (https://water.for-ident.org/). Note that log in credentials are needed.