Constructor and methods to handle non-target analysis results for mass spectrometry data

The MassSpecResults_NonTargetAnalysis class is a child of the Results class and is used to store results from non-target analysis (NTA) workflows for mass spectrometry data ("MassSpec").

Usage

MassSpecResults_NonTargetAnalysis(
  info = data.table::data.table(),
  headers = list(),
  features = list()
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
validate_object(x)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
show(x)

as.MassSpecResults_NonTargetAnalysis(value)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
x[i, j]

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
x[[value]]

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
get_features_count(x, analyses = NULL, filtered = FALSE)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
plot_features_count(
  x,
  analyses = NULL,
  filtered = FALSE,
  yLab = NULL,
  title = NULL,
  colorBy = "analyses",
  showLegend = TRUE,
  showHoverText = TRUE
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
get_features(
  x,
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  filtered = FALSE
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
map_features(
  x,
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  neutral_mass = TRUE,
  filtered = FALSE,
  legendNames = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  colorBy = "replicates+targets",
  showLegend = TRUE,
  interactive = TRUE,
  renderEngine = "webgl"
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
map_features_intensity(
  x,
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  filtered = FALSE,
  correctIntensity = FALSE,
  legendNames = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  colorBy = "replicates+targets",
  renderEngine = "webgl"
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
get_features_eic(
  x,
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  rtExpand = 0,
  mzExpand = 0,
  filtered = FALSE,
  useLoadedData = TRUE
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
plot_features(
  x,
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  rtExpand = 120,
  mzExpand = 0.001,
  useLoadedData = TRUE,
  filtered = FALSE,
  legendNames = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE,
  renderEngine = "webgl"
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
get_features_ms1(
  x,
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  rtWindow = c(-2, 2),
  mzWindow = c(-5, 100),
  mzClust = 0.003,
  presence = 0.8,
  minIntensity = 1000,
  normalized = TRUE,
  filtered = FALSE,
  useLoadedData = TRUE
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
plot_features_ms1(
  x,
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  rtWindow = c(-2, 2),
  mzWindow = c(-5, 100),
  mzClust = 0.003,
  presence = 0.8,
  minIntensity = 1000,
  normalized = TRUE,
  filtered = FALSE,
  useLoadedData = TRUE,
  legendNames = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  colorBy = "targets",
  showText = FALSE,
  interactive = TRUE
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
get_features_ms2(
  x,
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  isolationWindow = 1.3,
  mzClust = 0.003,
  presence = 0.8,
  minIntensity = 0,
  normalized = TRUE,
  filtered = FALSE,
  useLoadedData = TRUE
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
plot_features_ms2(
  x,
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  isolationWindow = 1.3,
  mzClust = 0.005,
  presence = 0.8,
  minIntensity = 0,
  normalized = TRUE,
  filtered = FALSE,
  useLoadedData = TRUE,
  legendNames = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  colorBy = "targets",
  showText = TRUE,
  interactive = TRUE
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
get_groups(
  x,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  filtered = FALSE,
  intensities = TRUE,
  average = FALSE,
  sdValues = FALSE,
  metadata = FALSE,
  correctIntensity = FALSE
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
plot_groups(
  x,
  analyses = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  rtExpand = 15,
  mzExpand = 0.001,
  filtered = FALSE,
  legendNames = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE,
  renderEngine = "webgl"
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
plot_groups_overview(
  x,
  analyses = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  rtExpand = 120,
  mzExpand = 0.005,
  useLoadedData = TRUE,
  correctIntensity = TRUE,
  filtered = FALSE,
  legendNames = NULL,
  title = NULL,
  heights = c(0.35, 0.5, 0.15),
  renderEngine = "webgl"
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
plot_groups_profile(
  x,
  analyses = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  filtered = FALSE,
  correctIntensity = TRUE,
  averaged = FALSE,
  normalized = TRUE,
  legendNames = NULL,
  yLab = NULL,
  title = NULL,
  showLegend = TRUE,
  renderEngine = "webgl"
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
get_groups_ms1(
  x,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  rtWindow = c(-2, 2),
  mzWindow = c(-5, 90),
  mzClustFeatures = 0.003,
  presenceFeatures = 0.8,
  minIntensityFeatures = 1000,
  useLoadedData = TRUE,
  mzClust = 0.003,
  presence = 0.8,
  minIntensity = 1000,
  top = 25,
  normalized = TRUE,
  groupBy = "groups",
  filtered = FALSE
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
get_groups_ms2(
  x,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  isolationWindow = 1.3,
  mzClustFeatures = 0.003,
  presenceFeatures = 0.8,
  minIntensityFeatures = 100,
  useLoadedData = TRUE,
  mzClust = 0.003,
  presence = 0.8,
  minIntensity = 100,
  top = 25,
  normalized = TRUE,
  groupBy = "groups",
  filtered = FALSE
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
plot_groups_ms1(
  x,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  rtWindow = c(-2, 2),
  mzWindow = c(-5, 90),
  mzClustFeatures = 0.005,
  presenceFeatures = 0.8,
  minIntensityFeatures = 1000,
  useLoadedData = TRUE,
  mzClust = 0.005,
  presence = 0.8,
  minIntensity = 1000,
  top = 25,
  normalized = TRUE,
  groupBy = "groups",
  filtered = FALSE,
  legendNames = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  colorBy = "targets",
  showText = FALSE,
  interactive = TRUE
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
plot_groups_ms2(
  x,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  isolationWindow = 1.3,
  mzClustFeatures = 0.003,
  presenceFeatures = 0.8,
  minIntensityFeatures = 100,
  useLoadedData = TRUE,
  mzClust = 0.003,
  presence = TRUE,
  minIntensity = 100,
  top = 25,
  normalized = TRUE,
  groupBy = "groups",
  filtered = FALSE,
  legendNames = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  colorBy = "targets",
  showText = TRUE,
  interactive = TRUE
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
get_components(
  x,
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  filtered = FALSE
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
map_components(
  x,
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  filtered = FALSE,
  legendNames = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE,
  showLegend = TRUE,
  renderEngine = "webgl"
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
get_suspects(
  x,
  analyses = NULL,
  database = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 5,
  sec = 10,
  millisec = 5,
  ppmMS2 = 10,
  mzrMS2 = 0.008,
  minCusiness = 0.7,
  minFragments = 3,
  filtered = FALSE
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
plot_suspects(
  x,
  analyses = NULL,
  database = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 5,
  sec = 10,
  millisec = 5,
  ppmMS2 = 10,
  mzrMS2 = 0.008,
  minCusiness = 0.7,
  minFragments = 3,
  filtered = FALSE,
  rtExpand = 120,
  mzExpand = 0.005,
  useLoadedData = TRUE,
  legendNames = NULL,
  colorBy = "replicates+targets",
  heights = c(0.5, 0.5),
  interactive = TRUE
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
get_internal_standards(x, average = TRUE)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
plot_internal_standards(
  x,
  analyses = NULL,
  showPresence = TRUE,
  showRecovery = TRUE,
  showDeviations = TRUE,
  showWidths = TRUE,
  renderEngine = "webgl"
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
get_compounds(
  x,
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 20,
  sec = 60,
  millisec = 5,
  filtered = FALSE,
  averaged = TRUE
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
get_fold_change(
  x,
  replicatesIn = NULL,
  replicatesOut = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 4,
  sec = 10,
  millisec = 5,
  filtered = FALSE,
  constantThreshold = 0.5,
  eliminationThreshold = 0.2,
  correctIntensity = FALSE,
  fillZerosWithLowerLimit = FALSE,
  lowerLimit = NA_real_
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
plot_fold_change(
  x,
  replicatesIn = NULL,
  replicatesOut = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  mobility = NULL,
  ppm = 4,
  sec = 10,
  millisec = 5,
  filtered = FALSE,
  constantThreshold = 0.5,
  eliminationThreshold = 0.2,
  correctIntensity = FALSE,
  fillZerosWithLowerLimit = FALSE,
  lowerLimit = NA_real_,
  normalized = TRUE,
  yLab = NULL,
  title = NULL,
  interactive = TRUE,
  showLegend = TRUE
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
get_patRoon_features(x, filtered = FALSE, featureGroups = TRUE)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
get_patRoon_MSPeakLists(
  x,
  mzClust = 0.005,
  minIntensity = 0,
  presence = 0.7,
  top = 25,
  normalized = FALSE
)

# S3 method for class 'MassSpecResults_NonTargetAnalysis'
get_patRoon_compounds(x)

Arguments

info

A data frame containing information about the analyses.

headers

A list of data frames containing information about the spectra headers.

features

A character vector with the feature IDs or a data.frame or data.table with a column named feature with the feature IDs. A column named analysis can be given to specify from which analyses the features matched.

x

A MassSpecResults_NonTargetAnalysis object.

value

An object that highly depends on the method. Check the specific method documentation for details.

i

A character or numeric vector to perform subsetting. Check the specific method for details.

j

A character or numeric vector to perform subsetting. Check the specific method for details.

analyses

Character or numeric vector with names or indexes of analyses in the Analyses object.

filtered

Logical (length 1). When TRUE, filtered features/groups are considered.

yLab

A string with the title for the y axis.

title

A string with the title.

colorBy

A string defining how to legend the plot. Possible values are analyses, targets (the default) or replicates.

showLegend

Logical (length 1). Set to TRUE to show legend.

showHoverText

Logical (length 1), set to TRUE to show hovering text annotations.

mass

A vector with target neutral mass value/s (in Da) or a two columns data.table or data.frame named min and max with minimum and maximum neutral mass values (in Da), respectively. Alternatively, neutral mass (in Da) and retention time (in seconds) and/or drift time values (in milliseconds) can be given as one data.table or data.frame with columns named mass and rt and/or drift. Then, the deviations given in the ppm, sec and millisec arguments are used to calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of neutral mass, retention time and drift time targets. In this case, the column names must be min, max, rtmin, rtmax, driftmin and driftmax. Note that when mass/time ranges are given, the ppm, sec and millisec arguments are not used.

mz

A vector with target m/z value/s (in Da) or a two columns data.table or data.frame named mzmin and mzmax with minimum and maximum m/z values (in Da), respectively. Alternatively, m/z (in Da) and retention time values (in seconds) can be given as one data.table or data.frame with columns named mz and rt and/or drift. Then, the deviations given in the ppm, sec and millisec arguments are used to calculate the ranges. Also works with a data.table or data.frame with minimum and maximum values of m/z, retention time and drift time targets. In this case, the column names must be mzmin, mzmax, rtmin, rtmax, driftmin and driftmax. Note that when mass/time ranges are given, the ppm, sec and millisec arguments are not used.

rt

A vector with target retention time values (in seconds) or a two columns data.table or data.frame with minimum and maximum retention time values (in seconds).

mobility

A vector with target drift time values (in milliseconds) or a two columns data.table or data.frame with minimum and maximum drift time values (in milliseconds).

ppm

Numeric of length one with the mass deviation, in ppm.

sec

Numeric of length one with the time deviation, in seconds.

millisec

Numeric of length one with the drift time deviation, in milliseconds.

neutral_mass

Logical, set to TRUE to return by neutral mass.

legendNames

A character vector with the same length as the targets or TRUE or FALSE for using the name in the added targets as legend of the plot.

xLab

A string with the title for the x axis.

interactive

Logical (length 1). When TRUE, the data is plotted interactively using plotly.

renderEngine

The engine to render the data. The default is "webgl".

correctIntensity

Logical (length 1) with TRUE for correcting the intensity for matrix suppression before applying intensity based filters. Note that suppression factor as obtained by the processing method CorrectMatrixSuppression must be present. When not available the correction is ignored even if set to TRUE.

rtExpand

Numeric (length 1) with the retention time values (in seconds) for expanding left and right the retention time range.

mzExpand

Numeric (length 1) with the mass or mass-to-charge ratio (m/z) (in Da) for expanding left and right the mass range.

useLoadedData

Logical of length one. Set to TRUE for using loaded data not raw data from files.

rtWindow

Numeric (length 2) with the retention time values (in seconds) for expanding left and right the retention time range. The first element expands left and the second expands right. Note that the first element should be negative to expand to the left.

mzWindow

Numeric (length 2) with the mass or mass-to-charge ratio (m/z) values (in Da) for expanding left and right the mass range. The first element expands left and the second expands right. Note that the first element should be negative to expand to the left.

mzClust

Numeric (length 1) with the mass deviation threshold (in Da) to cluster mass traces. Highly dependent on the mass resolution of the MS data.

presence

Numeric (length 1) with the required presence ratio from 0 (i.e., doesn't need to be present in any spectra) to 1 (i.e., must be present in all spectra) for traces during clustering of spectra.

minIntensity

Numeric (length 1) with the minimum intensity.

normalized

A logical value indicating whether the result should be normalized.

showText

Logical (length 1), set to TRUE to show the text annotations.

isolationWindow

Numeric value with the isolation window (in Da) applied for ion isolation and fragmentation during acquisition of tandem data (i.e., MS2 data).

groups

A numeric or character vector with the number or ID of feature groups, respectively.

intensities

Logical (length 1). When TRUE, feature intensity values are returned.

average

Logical (length 1). When TRUE, feature intensities from each group are averaged and returned for each analysis replicate group.

sdValues

Lofical of length 1. If TRUE, the standard deviation values will be included in the output.

metadata

Logical of length 1. If TRUE, the metadata will be included in the output.

heights

A numeric vector of length 2 or 3 to control the height of subplots. For plot_groups_overview the vector should have length 3 and for plot_suspects the vector should have length 2. The first value is the height of the first subplot, the second value is the height of the second subplot, and the third value (if present) is the height of the third subplot.

averaged

A logical value indicating whether the data is averaged.

mzClustFeatures

Numeric (length 1) with the mass deviation threshold (in Da) to cluster mass traces for features. Highly dependent on the mass resolution of the MS data.

presenceFeatures

Numeric (length 1) with the required presence ratio from 0 (i.e., doesn't need to be present in any spectra) to 1 (i.e., must be present in all spectra) for traces during clustering of spectra.

minIntensityFeatures

Numeric (length 1) with the minimum intensity for features.

top

Numeric of length 1 to set the top number of results to return.

groupBy

String with the way of grouping. Possible values are groups and replicates to group by feature groups or feature groups and replicates, respectively.

ppmMS2

Numeric (length 1) with the mass deviation for MS2, in ppm.

mzrMS2

Numeric (length 1). The m/z resolution for MS2 spectra.

minCusiness

Numeric (length 1). The minimum cusiness value for spectral comparison.

minFragments

Integer (length 1) with the minimum number of fragments.

showPresence

Logical (length 1). When TRUE the presence of the internal standards is plotted.

showRecovery

Logical (length 1). When TRUE the recovery of the internal standards is plotted.

showDeviations

Logical (length 1). When TRUE the deviations of the internal standards is plotted.

showWidths

Logical (length 1). When TRUE the widths of the internal standards is plotted.

replicatesIn

Character vector with the names of the replicates to be considered as the denominator.

replicatesOut

Character vector with the names of the replicates to be considered as the numerator.

constantThreshold

Numeric of length one. The threshold to consider a feature as constant.

eliminationThreshold

Numeric of length one. The threshold to consider a feature as eliminated.

fillZerosWithLowerLimit

Logical of length one. When TRUE the zero values are filled with the lower limit.

lowerLimit

Numeric of length one. The lower limit to fill the zero values.

featureGroups

Logical of length one. When TRUE the featureGroups class is returned.

Value

An object of class MassSpecResults_NonTargetAnalysis with the following structure:

type: The type of the results, which is "MassSpec".

name: The name of the results, which is "MassSpecResults_NonTargetAnalysis".

software: The software used for the analysis, which is "StreamFind".

version: The version of the software, as a character string.

info: A data frame containing information about the analyses.

headers: A list of data frames containing information about the spectra headers.

features: A list of data frames containing information about the features.

The info data.table contains the following columns: analysis, replicate, blank, polarity and file. Each features data frame contains the following columns: feature, group, rt, mz, intensity, area, rtmin, rtmax, mzmin, mzmax, mass, polarity, adduct, filtered, filter, filled, correction, eic, ms1, ms2, quality, annotation, istd, suspects, formulas and compounds.

Methods (by generic)

validate_object(MassSpecResults_NonTargetAnalysis): Validates the MassSpecResults_NonTargetAnalysis object, returning NULL if valid.

show(MassSpecResults_NonTargetAnalysis): Prints a summary of the MassSpecResults_NonTargetAnalysis object.

[: Subsets the MassSpecResults_NonTargetAnalysis object by analyses (with i) or feature groups (j).

[[: Subsets the MassSpecResults_NonTargetAnalysis object by feature groups. The argument value should be a character vector with the group names.

get_features_count(MassSpecResults_NonTargetAnalysis): Returns a data table with the number of features for each analysis.

plot_features_count(MassSpecResults_NonTargetAnalysis): Plots the number of features for each analysis as a bar plot.

get_features(MassSpecResults_NonTargetAnalysis): Returns a data table with the features for the specified analyses and targets.

map_features(MassSpecResults_NonTargetAnalysis): Maps features from the MassSpecResults_NonTargetAnalysis object.

map_features_intensity(MassSpecResults_NonTargetAnalysis): Maps features intensity from the MassSpecResults_NonTargetAnalysis object.

get_features_eic(MassSpecResults_NonTargetAnalysis): Returns a data table with the extracted ion chromatograms (EIC) for features in the specified analyses and targets.

plot_features(MassSpecResults_NonTargetAnalysis): Plots features from the MassSpecResults_NonTargetAnalysis object according to the specified parameters.

get_features_ms1(MassSpecResults_NonTargetAnalysis): Returns a data table with MS1 features for the specified analyses and targets.

plot_features_ms1(MassSpecResults_NonTargetAnalysis): Plots MS1 features from the MassSpecResults_NonTargetAnalysis object according to the specified parameters.

get_features_ms2(MassSpecResults_NonTargetAnalysis): Returns a data table with MS2 features for the specified analyses and targets.

plot_features_ms2(MassSpecResults_NonTargetAnalysis): Plots MS2 features from the MassSpecResults_NonTargetAnalysis object according to the specified parameters.

get_groups(MassSpecResults_NonTargetAnalysis): Returns a data table with group information for the specified analyses and targets.

plot_groups(MassSpecResults_NonTargetAnalysis): Plots groups from the MassSpecResults_NonTargetAnalysis object according to the specified parameters.

plot_groups_overview(MassSpecResults_NonTargetAnalysis): Plots an overview of groups from the MassSpecResults_NonTargetAnalysis object according to the specified parameters.

plot_groups_profile(MassSpecResults_NonTargetAnalysis): Plots the profile of groups from the MassSpecResults_NonTargetAnalysis object according to the specified parameters.

get_groups_ms1(MassSpecResults_NonTargetAnalysis): Extracts MS1 data for groups from the MassSpecResults_NonTargetAnalysis object according to the specified parameters.

get_groups_ms2(MassSpecResults_NonTargetAnalysis): Extracts MS2 data for groups from the MassSpecResults_NonTargetAnalysis object according to the specified parameters.

plot_groups_ms1(MassSpecResults_NonTargetAnalysis): Plots MS1 traces for groups from the MassSpecResults_NonTargetAnalysis object according to the specified parameters.

plot_groups_ms2(MassSpecResults_NonTargetAnalysis): Plots MS2 traces for groups from the MassSpecResults_NonTargetAnalysis object according to the specified parameters.

get_components(MassSpecResults_NonTargetAnalysis): Extracts components based on annotation from the MassSpecResults_NonTargetAnalysis object according to the specified parameters.

map_components(MassSpecResults_NonTargetAnalysis): Maps components based on annotation from the MassSpecResults_NonTargetAnalysis object according to the specified parameters.

get_suspects(MassSpecResults_NonTargetAnalysis): Extracts feature suspects based on a database. The database should be a data.frame with at least the columns name and mass, indicating the name and neutral monoisotopic mass of the suspect targets. If a database is not provided, it will extract suspects from the MassSpecResults_NonTargetAnalysis object if they were previously annotated.

plot_suspects(MassSpecResults_NonTargetAnalysis): Plots feature suspects based on a database. The database should be a data.frame with at least the columns name and mass, indicating the name and neutral monoisotopic mass of the suspect targets. If a database is not provided, it will extract suspects from the MassSpecResults_NonTargetAnalysis object if they were previously annotated.

get_internal_standards(MassSpecResults_NonTargetAnalysis): Extracts internal standards from the MassSpecResults_NonTargetAnalysis object. If the MassSpecResults_NonTargetAnalysis object has groups, it averages the internal standards across replicates, when average = TRUE.

plot_internal_standards(MassSpecResults_NonTargetAnalysis): Plots internal standards from the MassSpecResults_NonTargetAnalysis object. If the MassSpecResults_NonTargetAnalysis object has groups, it averages the internal standards across replicates, when average = TRUE.

get_compounds(MassSpecResults_NonTargetAnalysis): Extracts compounds from the MassSpecResults_NonTargetAnalysis object. If the MassSpecResults_NonTargetAnalysis object has groups, it averages the compounds across replicates, when averaged = TRUE.

get_fold_change(MassSpecResults_NonTargetAnalysis): Gets a data.table with fold-change analysis between the replicatesIn and replicatesOut.

plot_fold_change(MassSpecResults_NonTargetAnalysis): Plots the fold-change analysis between the replicatesIn and replicatesOut.

#' @param replicatesIn Character vector with the names of the replicates to be considered as the denominator.

get_patRoon_features(MassSpecResults_NonTargetAnalysis): Creates an S4 class features or featureGroupsfrom the patRoon package.

get_patRoon_MSPeakLists(MassSpecResults_NonTargetAnalysis): Creates S4 class MSPeakLists. Note that feature groups are required. The MS and MSMS spectra of each feature are then average by patRoon to produce the feature group spectra.

get_patRoon_compounds(MassSpecResults_NonTargetAnalysis): Creates an S4 class Compounds from the patRoon package.

Functions

as.MassSpecResults_NonTargetAnalysis(): Converts a list object to a MassSpecResults_NonTargetAnalysis object if it is compatible.