MassSpecSettings_BinSpectra_StreamFind

Bins spectral data according to units of a given variable (e.g., 5 retention time values) or based on bins given as a named list of numeric values, where the names are the bin labels (i.e. the name of the column) and the values are the bin dimensions (e.g. 5 seconds).

Usage

MassSpecSettings_BinSpectra_StreamFind(
  unitsVal = NULL,
  unitsNumber = NULL,
  bins = NULL,
  refBinAnalysis = NULL
)

Arguments

unitsVal

Character of length one with the column name of the variable to be used for binning.

unitsNumber

Integer of length one with the number of units to be used for binning.

bins

Named list of numeric values with the bin dimensions.

refBinAnalysis

The analysis index to use a reference for creating the bins.

Value

A MassSpecSettings_BinSpectra_StreamFind object.