MassSpecSettings_FindFeatures_xcms3_matchedfilter

Settings for finding features (i.e., chromatographic peaks) in mzML/mzXML files using the package xcms (version 3) with the algorithm MatchedFilter, which is optimal/preferred for low resolution LC-MS data. The function uses the package patRoon in the background.

Usage

MassSpecSettings_FindFeatures_xcms3_matchedfilter(
  binSize = 0.5,
  impute = "none",
  baseValue = 0,
  distance = 0,
  fwhm = 30,
  max = 5,
  snthresh = 20,
  steps = 2,
  mzdiff = 0.5,
  index = FALSE
)

Arguments

binSize

numeric(1) specifying the width of the bins/slices in m/z dimension.

impute

Character string specifying the method to be used for missing value imputation. Allowed values are "none" (no linear interpolation), "lin" (linear interpolation), "linbase" (linear interpolation within a certain bin-neighborhood) and "intlin".

baseValue

The base value to which empty elements should be set. This is only considered for impute as "linbase" and corresponds to the profBinLinBase's baselevel argument.

distance

For impute as "linbase": number of non-empty neighboring element of an empty element that should be considered for linear interpolation. See details section for more information.

fwhm

numeric(1) specifying the full width at half maximum of matched filtration gaussian model peak. Only used to calculate the actual sigma, see below.

max

numeric(1) representing the maximum number of peaks that are expected/will be identified per slice.

snthresh

numeric(1) defining the signal to noise ratio cutoff.

steps

numeric(1) defining the number of bins to be merged before filtration (i.e. the number of neighboring bins that will be joined to the slice in which filtration and peak detection will be performed).

mzdiff

numeric(1) representing the minimum difference in m/z dimension required for peaks with overlapping retention times; can be negative to allow overlap. During peak post-processing, peaks defined to be overlapping are reduced to the one peak with the largest signal.

index

logical(1) specifying whether indices should be returned instead of values for m/z and retention times.

Value

A MassSpecSettings_FindFeatures_xcms3_matchedfilter object.

Details

See the findFeaturesXCMS3 function from the patRoon package for more information and requirements.

References

Helmus R, ter Laak TL, van Wezel AP, de Voogt P, Schymanski EL (2021). “patRoon: open source software platform for environmental mass spectrometry based non-target screening.” Journal of Cheminformatics, 13(1). doi:10.1186/s13321-020-00477-w .

Helmus R, van de Velde B, Brunner AM, ter Laak TL, van Wezel AP, Schymanski EL (2022). “patRoon 2.0: Improved non-target analysis workflows including automated transformation product screening.” Journal of Open Source Software, 7(71), 4029. doi:10.21105/joss.04029 .

Smith, C.A., Want, E.J., O'Maille, G., Abagyan,R., Siuzdak, G. (2006). “XCMS: Processing mass spectrometry data for metabolite profiling using nonlinear peak alignment, matching and identification.” Analytical Chemistry, 78, 779–787.

Tautenhahn R, Boettcher C, Neumann S (2008). “Highly sensitive feature detection for high resolution LC/MS.” BMC Bioinformatics, 9, 504.

Benton HP, Want EJ, Ebbels TMD (2010). “Correction of mass calibration gaps in liquid chromatography-mass spectrometry metabolomics data.” BIOINFORMATICS, 26, 2488.