Results class child for Non-target Analysis Workflows for Mass Spectrometry Data

The NonTargetAnalysisResults class is a child of the Results class and is used to store results from non-target analysis (NTA) workflows for mass spectrometry data ("MassSpec"). It contains methods and properties specific to this type of analysis.

Usage

NonTargetAnalysisResults(
  analyses_info = data.table::data.table(),
  spectra_headers = list(),
  feature_list = list()
)

Arguments

analyses_info

A data frame containing information about the analyses.

spectra_headers

A list of data frames containing information about the spectra headers.

feature_list

A list of data frames containing information about the features.

Slots

analyses_info

(getter/setter) A data frame containing information about the analyses.

spectra_headers

(getter/setter) A list of data frames containing information about the spectra headers.

feature_list

(getter/setter) A list of data frames containing information about the features.

spectra_polarity

(getter) A character vector indicating the polarity of the spectra.

replicates

(getter/setter) A character vector indicating the replicates for each analysis.

blanks

(getter/setter) A character vector indicating the blanks for each analysis.

concentrations

(getter/setter) A numeric vector indicating the concentrations for each analysis.

number_analyses

(getter) An integer indicating the number of analyses.

number_features

(getter) An integer indicating the number of features.

has_features

(getter) A logical indicating if features are present.

has_filtered_features

(getter) A logical indicating if filtered features are present.

number_filtered_features

(getter) An integer indicating the number of filtered features.

has_groups

(getter) A logical indicating if groups are present.

number_groups

(getter) An integer indicating the number of groups.

group_names

(getter) A character vector indicating the names of the groups.

has_features_ms1

(getter) A logical indicating if MS1 features are present.

has_features_ms2

(getter) A logical indicating if MS2 features are present.

has_features_eic

(getter) A logical indicating if EIC features are present.

has_features_suspects

(getter) A logical indicating if suspect features are present.

info

(getter) A data frame containing an overview of the number of features.