RamanData R6 class and methods

(to be deprecate and replaced by RmanEngine) The RamanData R6 class is a framework with methods for parsing, processing, inspecting and storing RAMAN data.

Methods

Public methods

RamanData$new()

RamanData$print()

RamanData$get_headers()

RamanData$get_analyses()

RamanData$get_number_analyses()

RamanData$get_analysis_names()

RamanData$get_replicate_names()

RamanData$get_blank_names()

RamanData$get_files()

RamanData$get_spectra()

RamanData$get_overview()

RamanData$add_headers()

RamanData$add_analyses()

RamanData$add_replicate_names()

RamanData$add_blank_names()

RamanData$remove_analyses()

RamanData$merge_replicates()

RamanData$plot_spectra()

RamanData$clone()

Method new()

Creates an R6 RamanData class object. When headers are not given (i.e., NULL), a default Headers S3 class object is generated with name as NA_character, path as get_wd() and date as Sys.time(). See ?Headers for more information.

Usage

RamanData$new(files = NULL, headers = NULL, runParallel = FALSE)

Arguments

files

Full file paths of Raman analyses.

headers

A named list with headers or a ProjectHeaders S3 class object. Each list element must have length one. There is no type restrictions except for name (must be type character length 1), author (must be type character length 1), path (must be type character length 1) and date (must be class POSIXct or class POSIXt length 1). See more information in ?ProjectHeaders.

runParallel

Logical of length one. Set to TRUE for parallel processing. Note that parallel processing is not always faster due to overhead.

Method print()

Prints a summary of the RamanData object in the console.

Usage

RamanData$print()

Returns

Console text.

Method get_headers()

Gets the headers.

Usage

RamanData$get_headers(value = NULL)

Arguments

value

A character vector with the name/s of the header elements. When NULL, the entire headers list is returned.

Returns

The headers list or the header elements as defined by value.

Method get_analyses()

Gets analyses.

Usage

RamanData$get_analyses(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

The list of analyses or the analyses as defined by analyses argument.

Method get_number_analyses()

Gets the number of analyses present.

Usage

RamanData$get_number_analyses()

Returns

An integer value.

Method get_analysis_names()

Gets the analysis names.

Usage

RamanData$get_analysis_names(analyses = NULL)

Arguments

analyses

X.

Returns

A character vector.

Method get_replicate_names()

Gets the analysis replicate names.

Usage

RamanData$get_replicate_names(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

A character vector.

Method get_blank_names()

Gets the analysis blank replicate names.

Usage

RamanData$get_blank_names(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

A character vector.

Method get_files()

Gets the full file paths of the analyses.

Usage

RamanData$get_files(analyses = NULL)

Arguments

analyses

X.

Returns

A character vector.

Method get_spectra()

Gets spectra from analyses.

Usage

RamanData$get_spectra(
  analyses = NULL,
  rt = NULL,
  shift = NULL,
  minIntensity = 0
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

rt

Numeric of length two with the minimum and maximum retention time, in seconds.

shift

Numeric of length two with the minimum and maximum shift, in cm-1.

minIntensity

Numeric value with the minimum intensity.

Returns

A data.frame.

Method get_overview()

Gets the overview data.frame with all the analysis types, names, replicates, associated blank replicates and full file paths.

Usage

RamanData$get_overview()

Returns

A data.frame with columns type, analysis, replicate, blank and file.

Method add_headers()

Adds headers. If an argument or element "name" is given, it must be type character. If an argument or element path is given, it must be type character and exist. If an argument or element date is given, it must be class POSIXct or POSIXt. If given date is character, conversion to class POSIXct or POSIXt is attempted. See ?Headers for more information.

Usage

RamanData$add_headers(...)

Arguments

...

Arguments to be added as headers. Note that all given arguments must be of length one. There is no type limitations. The name of the argument becomes the name of the header entry. Alternatively, a single list argument with the headers (all length one) can be given.

Returns

Invisible.

Method add_analyses()

Adds analyses.

Usage

RamanData$add_analyses(analyses = NULL)

Arguments

analyses

A list.

Returns

Invisible.

Method add_replicate_names()

Adds or redefines the analysis replicate names.

Usage

RamanData$add_replicate_names(value = NULL)

Arguments

value

A character vector with the analysis replicate names. Must be of the same length as the number of analyses.

Returns

Invisible.

Method add_blank_names()

Adds or redefines the analysis blank replicate names.

Usage

RamanData$add_blank_names(value = NULL)

Arguments

value

A character vector with the analysis blank replicate names. Must be of the same length as the number of analyses.

Returns

Invisible.

Method remove_analyses()

Removes analyses.

Usage

RamanData$remove_analyses(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

Invisible.

Method merge_replicates()

Merges spectra for given RAMAN analyses from the same chromatographic separation when using LC-Raman coupling.

Usage

RamanData$merge_replicates(preCut = 2)

Arguments

preCut

The number of pre Raman scans to exclude when merging.

Returns

Invisible.

Method plot_spectra()

Plots spectra for given RAMAN analyses.

Usage

RamanData$plot_spectra(
  analyses = NULL,
  rt = NULL,
  shift = NULL,
  minIntensity = 0,
  xVal = "shift",
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  cex = 0.6,
  showLegend = TRUE,
  colorBy = "analyses",
  interactive = TRUE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

rt

Numeric of length two with the minimum and maximum retention time, in seconds.

shift

Numeric of length two with the minimum and maximum shift, in cm-1.

minIntensity

Numeric value with the minimum intensity.

xVal

Character of length one. Possible are "rt" or "shift" for using the retention time or the shift as x axis, respectively.

xLab

A string with the title for the x axis.

yLab

A string with the title for the y axis.

title

A string with the title.

cex

A length one numeric value defining the scale of the legend for non interactive plotting.

showLegend

Logical. Set to TRUE to show legend.

colorBy

A string of length 1. One of analyses (the default) or replicates.

interactive

Logical. When TRUE, the data is plotted interactively using plotly.

Returns

A plot.

Method clone()

The objects of this class are cloneable with this method.

Usage

RamanData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.