Raman Method for Adding Shift Values to Spectra (native algorithm)

Adds custom shift values to the Raman spectra or copies from one or more analyses to the other analyses. When more than one analyses are selected, the shift values are copied from the selected but them averaged before being added to the other analyses. If the column "rt" is present in the spectra, the shift values are added to each retention time.

Usage

RamanMethod_AddShiftValues_native(
  mode = "analyses",
  index = NA_integer_,
  shifts = NA_real_
)

Arguments

mode

A character string. The mode to be used, possible values are analyses or custom for copying from analyses or adding custom shift values, respectively.

index

An integer vector with the index/es of the analysis/analyses to copy the shift values from.

shifts

A numeric vector with the shift values to be added to the spectra.

Value

A RamanMethod_AddShiftValues_native object.