Raman Method for Binning Time Scans (native algorithm)

Merges scans by averaging spectra according to a given binning size.

Usage

RamanMethod_BinScans_native(
  mode = c("unit"),
  value = 5,
  global = FALSE,
  refAnalysis = NA_integer_
)

Arguments

mode

Character (length 1) with the binning mode. Possible values are scans and time. When scans is selected, the binning is done by the number of scans. When time is selected, the binning is done by time unit (e.g., seconds or minutes).

value

Numeric (length 1) with the binning size.

global

Logical (length 1) to perform a global binning. If TRUE, the binning is performed over all analyses. If FALSE, the binning is performed for each analysis.

refAnalysis

Integer (length 1) with the reference analysis index to use for binning.

Value

A RamanMethod_BinScans_native object.