Generic Results Class

The Results class is used to store results of data processing in Analyses child classes. Child classes of Results are implemented for diverse types of results for a given type of data.

Usage

Results(
  type = NA_character_,
  name = "Results",
  software = "StreamFind",
  version = NA_character_
)

# S3 method for class 'Results'
validate_object(x)

Arguments

type

A character string indicating the type of data.

name

A character string representing the name of the results.

software

A character string representing the name of the software used to generate the results.

version

A character string representing the version of the software used to generate the results.

x

An object of class Results.

Value

A Results S3 class object which is a list with the elements type, name, software, and version. Other elements are added by child class constructors (e.g. MassSpecSpectra).

Methods (by generic)

validate_object(Results): Validate the Results object, returning NULL if valid.