Settings_filter_features_patRoon

Settings for filtering of features and feature groups. A full description of the filtering parameters is in replicateGroupSubtract from patRoon package.

Usage

Settings_filter_features_patRoon(
  absMinIntensity = NULL,
  relMinIntensity = NULL,
  preAbsMinIntensity = NULL,
  preRelMinIntensity = NULL,
  absMinAnalyses = NULL,
  relMinAnalyses = NULL,
  absMinReplicates = NULL,
  relMinReplicates = NULL,
  absMinFeatures = NULL,
  relMinFeatures = NULL,
  absMinReplicateAbundance = NULL,
  relMinReplicateAbundance = NULL,
  absMinConc = NULL,
  relMinConc = NULL,
  absMaxTox = NULL,
  relMaxTox = NULL,
  absMinConcTox = NULL,
  relMinConcTox = NULL,
  maxReplicateIntRSD = NULL,
  blankThreshold = NULL,
  retentionRange = NULL,
  mzRange = NULL,
  mzDefectRange = NULL,
  chromWidthRange = NULL,
  featQualityRange = NULL,
  groupQualityRange = NULL,
  rGroups = NULL,
  results = NULL,
  removeBlanks = FALSE,
  removeISTDs = FALSE,
  checkFeaturesSession = NULL,
  removeNA = FALSE,
  negate = FALSE
)

Arguments

absMinIntensity

Numeric length one. Minimum absolute intensity for a feature.

relMinIntensity

Numeric length one. Minimum relative intensity for a feature.

preAbsMinIntensity

Numeric length one. Minimum absolute intensity for a feature before grouping.

preRelMinIntensity

Numeric length one. Minimum relative intensity for a feature before grouping.

absMinAnalyses

Numeric length one. Minimum number of analyses a feature must be present in.

relMinAnalyses

Numeric length one. Minimum relative number of analyses a feature must be present in.

absMinReplicates

Numeric length one. Minimum number of replicates a feature must be present in.

relMinReplicates

Numeric length one. Minimum relative number of replicates a feature must be present in.

absMinFeatures

Numeric length one. Minimum number of features a feature group must contain.

relMinFeatures

Numeric length one. Minimum relative number of features a feature group must contain.

absMinReplicateAbundance

Numeric length one. Minimum absolute abundance of a replicate.

relMinReplicateAbundance

Numeric length one. Minimum relative abundance of a replicate.

absMinConc

Numeric length one. Minimum absolute concentration of a feature.

relMinConc

Numeric length one. Minimum relative concentration of a feature.

absMaxTox

Numeric length one. Maximum absolute toxicity of a feature.

relMaxTox

Numeric length one. Maximum relative toxicity of a feature.

absMinConcTox

Numeric length one. Minimum absolute concentration of a feature to be considered toxic.

relMinConcTox

Numeric length one. Minimum relative concentration of a feature to be considered toxic.

maxReplicateIntRSD

Numeric length one. Maximum relative standard deviation of intensities within a replicate.

blankThreshold

Numeric length one. Maximum intensity of a feature to be considered a blank.

retentionRange

Numeric length two. Retention time range (in seconds) for a feature.

mzRange

Numeric length two. m/z range (in Da) for a feature.

mzDefectRange

Numeric length two. m/z defect range (in Da) for a feature.

chromWidthRange

Numeric length two. Chromatographic width range (in seconds) for a feature.

featQualityRange

Numeric length two. Feature quality range for a feature.

groupQualityRange

Numeric length two. Group quality range for a feature group.

rGroups

List of replicate groups.

results

Only keep feature groups that have results in the object specified by results. See replicateGroupSubtract for further information.

removeBlanks

Logical length one. Remove blank samples.

removeISTDs

Logical length one. Remove internal standards.

checkFeaturesSession

Check features session.

removeNA

Logical length one. Remove NA values.

negate

Logical length one. Negate the filter.

Value

A ProcessingSettings S3 class object with subclass Settings_filter_features_patRoon.

Details

Note that when filters are applied to features or feature groups these require specific results from processing modules. For instance, subtracting the blank can only be done after grouping features. Also, some filters require. Thus, not all filters can be applied to features. See features-class and replicateGroupSubtract for further information.