Settings_find_features_kpic2

Settings for finding features (i.e., chromatographic peaks) in mzML/mzXML files using the package KPIC. The function uses the package patRoon in the background.

Usage

Settings_find_features_kpic2(
  level = 500,
  mztol = 0.01,
  gap = 2,
  width = 5,
  min_snr = 4,
  kmeans = TRUE,
  alpha = 0.3
)

# S3 method for Settings_find_features_kpic2
validate(x)

Arguments

level

Mass traces are only retained if their maximum values are over level.

mztol

The initial m/z tolerance.

gap

The number of gap points of a mass trace.

width

The minimum length of a mass trace.

min_snr

Minimum signal to noise ratio.

kmeans

If TRUE, getPIC.kmeans is used to obtain PICs (i.e., features). If FALSE, getPIC is used.

alpha

If kmeans is TRUE, alpha is the parameter of forecasting. If kmeans is FALSE, alpha is not used.

x

A Settings_find_features_kpic2 S3 class object.

Value

A ProcessingSettings S3 class object with subclass Settings_find_features_kpic2.

Details

See the findFeaturesKPIC2 function from the patRoon package for more information and requirements.

Methods (by generic)

validate(Settings_find_features_kpic2): Validates the object structure, returning a logical value of length one.

References

Helmus R, ter Laak TL, van Wezel AP, de Voogt P, Schymanski EL (2021). “patRoon: open source software platform for environmental mass spectrometry based non-target screening.” Journal of Cheminformatics, 13(1). doi:10.1186/s13321-020-00477-w .

Helmus R, van de Velde B, Brunner AM, ter Laak TL, van Wezel AP, Schymanski EL (2022). “patRoon 2.0: Improved non-target analysis workflows including automated transformation product screening.” Journal of Open Source Software, 7(71), 4029. doi:10.21105/joss.04029 .

Ji H, Zeng F, Xu Y, Lu H, Zhang Z (2017). “KPIC2: An Effective Framework for Mass Spectrometry-Based Metabolomics Using Pure Ion Chromatograms.” Anal Chem., 14(89), 7631--7640. doi:10.1021/acs.analchem.7b01547 , https://pubs.acs.org/doi/10.1021/acs.analchem.7b01547.