Settings_generate_compounds_metfrag

Settings for generating compounds using MetFrag. The algorithm is used via the function generateCompounds from the package patRoon. Therefore, it is highly recommended to check the original documentation of the function in patRoon for more details.

Usage

Settings_generate_compounds_metfrag(
  method = "CL",
  timeout = 300,
  timeoutRetries = 5,
  errorRetries = 5,
  topMost = 5,
  dbRelMzDev = 8,
  fragRelMzDev = 10,
  fragAbsMzDev = 0.005,
  adduct = NULL,
  database = "comptox",
  extendedPubChem = "auto",
  chemSpiderToken = "",
  scoreTypes = patRoon::compoundScorings("metfrag", "comptox", onlyDefault = TRUE)$name,
  scoreWeights = 1,
  preProcessingFilters = c("UnconnectedCompoundFilter", "IsotopeFilter"),
  postProcessingFilters = c("InChIKeyFilter"),
  maxCandidatesToStop = 100,
  identifiers = NULL,
  extraOpts = NULL
)

Arguments

method

Character (length 1) with the method to be used for MetFrag execution: "CL" for MetFragCL and "R" for MetFragR.

timeout

Numeric (length 1) with the maximum time (in seconds) before a MetFrag query for a feature group is stopped.

timeoutRetries

Numeric (length 1) with the maximum number of retries after reaching a timeout before completely skipping the MetFrag query for a feature group.

errorRetries

Numeric (length 1) with the maximum number of retries after an error occurred.

topMost

Numeric (length 1) with the maximum number of top candidates to be returned.

dbRelMzDev

Numeric (length 1) with the relative mass deviation, in ppm, for the database search.

fragRelMzDev

Numeric (length 1) with the relative mass deviation, in ppm, for the fragment search.

fragAbsMzDev

Numeric (length 1) with the absolute mass deviation, in Da, for the fragment search.

adduct

Character (length 1) with the adduct to be used for the MetFrag annotation.

database

Character (length 1) with the database to be used for the MetFrag annotation. Valid values are: "pubchem", "chemspider", "for-ident", "comptox", "pubchemlite", "kegg", "sdf", "psv" and "csv".

extendedPubChem

Extended PubChem database is used for the MetFrag annotation when database is "pubchem". Valid values are: FALSE (never use it), TRUE (always use it) or "auto" (default, use if specified scorings demand it).

chemSpiderToken

Character (length 1) with the ChemSpider token to be used for the MetFrag annotation when database is "chemspider".

scoreTypes

Character vector with the score types to be used for the MetFrag annotation.

scoreWeights

Numeric vector with the score weights to be used for the MetFrag annotation.

preProcessingFilters

Character vector with the pre-processing filters to be used for the MetFrag annotation.

postProcessingFilters

Character vector with the post-processing filters to be used for the MetFrag annotation.

maxCandidatesToStop

Numeric (length 1) with the maximum number of candidates to be returned before stopping the MetFrag query for a feature group.

identifiers

A list containing for each feature group a character vector with database identifiers that should be used to find candidates for a feature group (the list should be named by feature group names). Can be NULL.

extraOpts

A named list containing further settings MetFrag.

Value

A ProcessingSettings S3 class object with subclass Settings_generate_compounds_metfrag.

Details

Detailed documentation can be found in generateCompoundsMetFrag.

References

Ruttkies C, Schymanski EL, Wolf S, Hollender J, Neumann S (2016). “MetFrag relaunched: incorporating strategies beyond in silico fragmentation.” Journal of cheminformatics, 8(1), 1--16.

Ruttkies C, Neumann S, Posch S (2019). “Improving MetFrag with statistical learning of fragment annotations.” BMC bioinformatics, 20(1), 1--14.

Ruttkies C, Schymanski EL, Strehmel N, Hollender J, Neumann S, Williams AJ, Krauss M (2019). “Supporting non-target identification by adding hydrogen deuterium exchange MS/MS capabilities to MetFrag.” Analytical and bioanalytical chemistry, 411, 4683--4700.

Wolf S, Schmidt S, Müller-Hannemann M, Neumann S (2010). “In silico fragmentation for computer assisted identification of metabolite mass spectra.” BMC bioinformatics, 11, 1--12.

Helmus R, ter Laak TL, van Wezel AP, de Voogt P, Schymanski EL (2021). “patRoon: open source software platform for environmental mass spectrometry based non-target screening.” Journal of Cheminformatics, 13(1). doi:10.1186/s13321-020-00477-w .

Helmus R, van de Velde B, Brunner AM, ter Laak TL, van Wezel AP, Schymanski EL (2022). “patRoon 2.0: Improved non-target analysis workflows including automated transformation product screening.” Journal of Open Source Software, 7(71), 4029. doi:10.21105/joss.04029 .