Settings_group_features_openms

Settings for grouping features (i.e., chromatographic peaks) in mzML/mzXML files using the OpenMS(https://abibuilder.cs.uni-tuebingen.de/archive/openms/) software with the algorithm FeatureLinkerUnlabeled. The function uses the package patRoon in the background.

Usage

Settings_group_features_openms(
  rtalign = FALSE,
  QT = FALSE,
  maxAlignRT = 5,
  maxAlignMZ = 0.008,
  maxGroupRT = 5,
  maxGroupMZ = 0.008,
  extraOptsRT = NULL,
  extraOptsGroup = NULL,
  verbose = FALSE
)

Arguments

rtalign

Logical length one. Set to TRUE to enable retention time alignment.

QT

Logical length one. When TRUE the FeatureLinkerUnlabeledQT is used instead of FeatureLinkerUnlabeled for grouping features.

maxAlignRT

Numeric length one. Maximum retention time (in seconds) for feature pairing when performing retention time alignment.

maxAlignMZ

Numeric length one. Maximum m/z (in Da) for feature pairing when performing retention time alignment.

maxGroupRT

Numeric length one. Maximum retention time (in seconds) for feature pairing when performing grouping.

maxGroupMZ

Numeric length one. Maximum m/z (in Da) for feature pairing when performing grouping.

extraOptsRT

Named list containing extra options that will be passed to MapAlignerPoseClustering.

extraOptsGroup

Named list containing extra options that will be passed to FeatureLinkerUnlabeledQT/FeatureLinkerUnlabeled.

verbose

Logical of length one. When TRUE adds processing information to the console.

Value

A ProcessingSettings S3 class object with subclass Settings_group_features_openms.

Details

See the groupFeaturesOpenMS function from the patRoon package for more information and requirements.

References

Helmus R, ter Laak TL, van Wezel AP, de Voogt P, Schymanski EL (2021). “patRoon: open source software platform for environmental mass spectrometry based non-target screening.” Journal of Cheminformatics, 13(1). doi:10.1186/s13321-020-00477-w .

Helmus R, van de Velde B, Brunner AM, ter Laak TL, van Wezel AP, Schymanski EL (2022). “patRoon 2.0: Improved non-target analysis workflows including automated transformation product screening.” Journal of Open Source Software, 7(71), 4029. doi:10.21105/joss.04029 .

Röst HL, Sachsenberg T, Aiche S, Bielow C, Weisser H, Aicheler F, Andreotti S, Ehrlich H, Gutenbrunner P, Kenar E, Liang X, Nahnsen S, Nilse L, Pfeuffer J, Rosenberger G, Rurik M, Schmitt U, Veit J, Walzer M, Wojnar D, Wolski WE, Schilling O, Choudhary JS, Malmström L, Aebersold R, Reinert K, Kohlbacher O (2016). “OpenMS: a flexible open-source software platform for mass spectrometry data analysis.” Nature Methods, 13(9), 741--748. doi:10.1038/nmeth.3959 .