Settings_group_features_xcms3_peakdensity_peakgroups

Settings for aligning and grouping features (i.e., chromatographic peaks) across mzML/mzXML files using the package xcms (version 3) with the algorithm peakGroups for retention time alignment and the algorithm peakdensity for grouping. The function uses the package patRoon in the background.

Usage

Settings_group_features_xcms3_peakdensity_peakgroups(
  bw = 5,
  minFraction = 1,
  minSamples = 1,
  binSize = 0.008,
  pre_bw = 5,
  pre_minFraction = 1,
  pre_minSamples = 1,
  pre_binSize = 0.008,
  maxFeatures = 100,
  rtAlignMinFraction = 0.9,
  extraPeaks = 1,
  smooth = "loess",
  span = 0.2,
  family = "gaussian",
  peakGroupsMatrix = matrix(nrow = 0, ncol = 0),
  subset = integer(),
  subsetAdjust = "average"
)

# S3 method for Settings_group_features_xcms3_peakdensity_peakgroups
validate(x)

Arguments

bw

numeric(1) defining the bandwidth (standard deviation of the smoothing kernel) to be used. This argument is passed to the density() method.

minFraction

numeric(1) defining the minimum fraction of analyses in at least one analysis replicate group in which the features have to be present to be considered as a feature group.

minSamples

numeric(1) with the minimum number of analyses in at least one analysis replicate group in which the features have to be detected to be considered a feature group.

binSize

numeric(1) defining the size of the overlapping slices in mz dimension.

pre_bw

as bw but applied before retention time alignment.

pre_minFraction

as minFraction but applied before retention time alignment.

pre_minSamples

as minSamples but applied before retention time alignment.

pre_binSize

as binSize but applied before retention time alignment.

maxFeatures

numeric(1) with the maximum number of feature groups to be identified in a single mz slice.

rtAlignMinFraction

numeric(1) between 0 and 1 defining the minimum required fraction of samples in which peaks for the peak group were identified. Peak groups passing this criteria will aligned across samples and retention times of individual spectra will be adjusted based on this alignment. For minFraction = 1 the peak group has to contain peaks in all samples of the experiment. Note that if subset is provided, the specified fraction is relative to the defined subset of samples and not to the total number of samples within the experiment (i.e. a peak has to be present in the specified proportion of subset samples).

extraPeaks

numeric(1) defining the maximal number of additional peaks for all samples to be assigned to a peak group (i.e. feature) for retention time correction. For a data set with 6 samples, extraPeaks = 1 uses all peak groups with a total peak count <= 6 + 1. The total peak count is the total number of peaks being assigned to a peak group and considers also multiple peaks within a sample being assigned to the group.

smooth

character defining the function to be used, to interpolate corrected retention times for all peak groups. Either "loess" or "linear".

span

numeric(1) defining the degree of smoothing (if smooth = "loess"). This parameter is passed to the internal call to loess.

family

character defining the method to be used for loess smoothing. Allowed values are "gaussian" and "symmetric".See loess for more information.

peakGroupsMatrix

optional matrix of (raw) retention times for the peak groups on which the alignment should be performed. Each column represents a sample, each row a feature/peak group. Such a matrix is for example returned by the adjustRtimePeakGroups method.

subset

integer with the indices of samples within the experiment on which the alignment models should be estimated. Samples not part of the subset are adjusted based on the closest subset sample. See description above for more details.

subsetAdjust

character specifying the method with which non-subset samples should be adjusted. Supported options are "previous" and "average" (default). See description above for more information.

x

A Settings_group_features_xcms3_peakdensity_peakgroups S3 class object.

Value

A ProcessingSettings S3 class object with subclass Settings_group_features_xcms3_peakdensity_peakgroups.

Details

See the groupFeaturesXCMS3 function from the patRoon package for more information and requirements.

Methods (by generic)

validate(Settings_group_features_xcms3_peakdensity_peakgroups): Validates the object structure, returning a logical value of length one.

References

Helmus R, ter Laak TL, van Wezel AP, de Voogt P, Schymanski EL (2021). “patRoon: open source software platform for environmental mass spectrometry based non-target screening.” Journal of Cheminformatics, 13(1). doi:10.1186/s13321-020-00477-w .

Helmus R, van de Velde B, Brunner AM, ter Laak TL, van Wezel AP, Schymanski EL (2022). “patRoon 2.0: Improved non-target analysis workflows including automated transformation product screening.” Journal of Open Source Software, 7(71), 4029. doi:10.21105/joss.04029 .

Smith, C.A., Want, E.J., O'Maille, G., Abagyan,R., Siuzdak, G. (2006). “XCMS: Processing mass spectrometry data for metabolite profiling using nonlinear peak alignment, matching and identification.” Analytical Chemistry, 78, 779--787.

Tautenhahn R, Boettcher C, Neumann S (2008). “Highly sensitive feature detection for high resolution LC/MS.” BMC Bioinformatics, 9, 504.

Benton HP, Want EJ, Ebbels TMD (2010). “Correction of mass calibration gaps in liquid chromatography-mass spectrometry metabolomics data.” BIOINFORMATICS, 26, 2488.