StatisticMethod_MakeModel_mcrals_mdatools S7 class

Makes a Multivariate Curve Resolution (MCR) model using Alternating Least Squares (ALS) based on the R package mdatools.

Usage

StatisticMethod_MakeModel_mcrals_mdatools(
  ncomp = NULL,
  cont.solver = "mcrals.nnls",
  spec.solver = "mcrals.nnls",
  exclrows = NULL,
  exclcols = NULL,
  verbose = FALSE,
  max.niter = 100,
  tol = 10^-6,
  info = ""
)

Arguments

ncomp

Integer (length 1) with the number of components to be calculated.

cont.solver

Character (length 1) with the name of the to solve the pure components contributions (see details in mcrals).

spec.solver

Character (length 1) with the name of the function to solve the pure components spectra (see details in mcrals).

exclrows

Integer vector with the row indices to be excluded.

exclcols

Integer vector with the column indices to be excluded.

verbose

Logical (length 1) indicating if the function should be verbose.

max.niter

Integer (length 1) with the maximum number of iterations.

tol

Numeric (length 1) with the tolerance for convergence.

info

Character (length 1) with additional information.

Value

A StatisticMethod_MakeModel_mcrals_mdatools object.

Note

The functionality to define the initial pure components spectra as well as forcing concentration and spectra values are not included in this integration of mcrals.

References

Kucheryavskiy S (2020). “mdatools – R package for chemometrics.” Chemometrics and Intelligent Laboratory Systems, 198, 103937. ISSN 0169-7439, doi:10.1016/j.chemolab.2020.103937 , https://www.sciencedirect.com/science/article/pii/S0169743919305672.

Windig W, Gallagher NB, Shaver JM, Wise BM (2005). “A new approach for interactive self-modeling mixture analysis.” Chemometrics and Intelligent Laboratory Systems, 77(1), 85-96. ISSN 0169-7439, doi:10.1016/j.chemolab.2004.06.009 , FESTSCHRIFT HONOURING PROFESSOR D.L. MASSART, https://www.sciencedirect.com/science/article/pii/S0169743904002011.