StatisticMethod_MakeModel_pca_mdatools S7 class

Makes a Principle Component Analysis (PCA) model based on the R package mdatools.

Usage

StatisticMethod_MakeModel_pca_mdatools(
  center = FALSE,
  scale = FALSE,
  ncomp = NULL,
  exclrows = NULL,
  exclcols = NULL,
  x.test = NULL,
  method = "svd",
  rand = NULL,
  lim.type = "ddmoments",
  alpha = 0.05,
  gamma = 0.01,
  info = ""
)

Arguments

center

Logical (length 1) indicating if the data should be centered.

scale

Logical (length 1) indicating if the data should be scaled.

ncomp

Integer (length 1) with the number of components to be calculated.

exclrows

Integer vector with the row indices to be excluded.

exclcols

Integer vector with the column indices to be excluded.

x.test

A data.frame with the test data.

method

Character (length 1) with the method to be used for PCA. Possible values are "svd" and "nipals".

rand

Integer (length 1) with the random seed.

lim.type

Character (length 1) with the type of limit for the PCA. Possible values are "jm" Jackson-Mudholkar approach approach, "chisq" based on chi-square distribution, "ddmoments" and "ddrobust" related to data driven method. It is highly recommended to concult the original documentation in pca for more information.

alpha

Numeric (length 1) with the alpha value for the PCA.

gamma

Numeric (length 1) with the gamma value for the PCA.

info

Character (length 1) with additional information.

Value

A StatisticMethod_MakeModel_pca_mdatools object.

References

mdatools01StreamFind