StatisticResults_MCRALS_mdatools Class

Class for storing results of MCR-ALS models created with the mdatools package.

Usage

StatisticResults_MCRALS_mdatools(model = list())

# S3 method for class 'StatisticResults_MCRALS_mdatools'
validate_object(x)

# S3 method for class 'StatisticResults_MCRALS_mdatools'
summary(object, ...)

# S3 method for class 'StatisticResults_MCRALS_mdatools'
predict(x, data)

# S3 method for class 'StatisticResults_MCRALS_mdatools'
test(x, data)

# S3 method for class 'StatisticResults_MCRALS_mdatools'
get_model_data(x)

# S3 method for class 'StatisticResults_MCRALS_mdatools'
plot_explained_variance(
  x,
  interactive = TRUE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  showText = TRUE,
  showLegend = TRUE
)

# S3 method for class 'StatisticResults_MCRALS_mdatools'
plot_cumulative_explained_variance(
  x,
  interactive = TRUE,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  showText = TRUE,
  showLegend = TRUE
)

# S3 method for class 'StatisticResults_MCRALS_mdatools'
plot_resolved_spectra(
  x,
  interactive = TRUE,
  pcs = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  showText = TRUE,
  showLegend = TRUE
)

# S3 method for class 'StatisticResults_MCRALS_mdatools'
plot_contributions(
  x,
  interactive = TRUE,
  pcs = NULL,
  xLab = NULL,
  yLab = NULL,
  title = NULL
)

# S3 method for class 'StatisticResults_MCRALS_mdatools'
plot_overview(x, ...)

Arguments

model

A mcrals object created with the mdatools package.

x

A StatisticResults_MCRALS_mdatools object.

object

A StatisticResults_MCRALS_mdatools object.

...

Additional arguments passed to the plot function.

data

A data frame containing the new data to test or predict.

interactive

Logical (length 1). When TRUE, the data is plotted interactively using plotly.

xLab

A string with the title for the x axis.

yLab

A string with the title for the y axis.

title

A string with the title.

showText

Logical (length 1), set to TRUE to show the text annotations.

showLegend

Logical (length 1). Set to TRUE to show legend.

pcs

A vector of integers specifying the principal components to plot. If NULL, all components are plotted.

Methods (by generic)

validate_object(StatisticResults_MCRALS_mdatools): Validate the object, return NULL if valid.

summary(StatisticResults_MCRALS_mdatools): Print a summary of the model.

predict(StatisticResults_MCRALS_mdatools): Predict new data using the MCRALS model.

test(StatisticResults_MCRALS_mdatools): Test new data using the MCRALS model.

get_model_data(StatisticResults_MCRALS_mdatools): Get the model data as a list of data frames.

plot_explained_variance(StatisticResults_MCRALS_mdatools): Plot the explained variance of the model.

plot_cumulative_explained_variance(StatisticResults_MCRALS_mdatools): Plot the cumulative explained variance of the model.

plot_resolved_spectra(StatisticResults_MCRALS_mdatools): Plot the resolved spectra of the MCRALS model.

plot_contributions(StatisticResults_MCRALS_mdatools): Plot the contributions of the resolved spectra.

plot_overview(StatisticResults_MCRALS_mdatools): Plot the explained variance of the MCRALS model.