StreamFind: Framework and workflows for data analysis in the water cycle

The StreamFind R package is a framework for spectrometric and spectroscopic data processing with focus on environmental and quality studies of the water cycle. The StreamFind R package aims to stimulate the use of advanced data analysis (e.g., non-target screening (NTS) and statistical analysis) in routine studies, promoting standardization of data processing and easing the retrospective inclusion of data.

See also

Author

Maintainer: Ricardo Cunha cunha@iuta.de (ORCID)

Other contributors:

Walter Laurito laurito@fzi.de [contributor]