Generic (top level) Workflow class and methods

The Workflow class is an ordered list of ProcessingStep objects dedicated to a specific type of data.

Usage

Workflow(processing_steps = list())

# S3 method for class 'Workflow'
validate_object(x)

# S3 method for class 'Workflow'
info(x)

# S3 method for class 'Workflow'
get_methods(x)

# S3 method for class 'Workflow'
x[i]

# S3 method for class 'Workflow'
x[i] <- value

# S3 method for class 'Workflow'
x[[i]]

# S3 method for class 'Workflow'
x[[i]] <- value

# S3 method for class 'Workflow'
save(x, file = "workflow.rds")

# S3 method for class 'Workflow'
read(x, file)

# S3 method for class 'Workflow'
show(x, ...)

Arguments

processing_steps

A list of ProcessingStep objects.

x

A Workflow object.

i

An index indicating which element to replace or a character to add a new element.

value

A ProcessingStep object to replace the selected element with.

file

A file path to read the Workflow object from. Supported formats are .json and .rds.

...

Additional arguments (not used).

Methods (by generic)

validate_object(Workflow): Validate a Workflow object, returning NULL if valid or an error if not.

info(Workflow): Convert a Workflow object to a data frame.

get_methods(Workflow): Get the names of the methods in a Workflow object.

[: Subset a Workflow object.

`[`(Workflow) <- value: Replace elements in a Workflow object, returning a modified Workflow object.

[[: Extract elements from a Workflow object using [[.

`[[`(Workflow) <- value: Replace elements in a Workflow object using [[<-.

save(Workflow): Save a Workflow object to a file.

read(Workflow): Read a Workflow object from a file.

show(Workflow): Show the contents of a Workflow object.