msAnalysis S3 class constructor and methods

Creates an msAnalysis S3 class object.

Usage

msAnalysis(
  name = NA_character_,
  replicate = NA_character_,
  blank = NA_character_,
  file = NA_character_,
  type = NA_character_,
  instrument = list(),
  time_stamp = NA_character_,
  spectra_number = NA_integer_,
  spectra_mode = NA_character_,
  spectra_levels = NA_integer_,
  mz_low = NA_real_,
  mz_high = NA_real_,
  rt_start = NA_real_,
  rt_end = NA_real_,
  polarity = NA_character_,
  chromatograms_number = NA_integer_,
  ion_mobility = NA,
  tic = data.table(),
  bpc = data.table(),
  spectra = data.table(),
  chromatograms = data.table(),
  features = data.table(),
  metadata = list()
)

# S3 method for msAnalysis
validate(msAnalysis = NULL)

as.msAnalysis(value)

parse.msAnalysis(files = NULL, runParallel = FALSE)

Arguments

name

mzML or mzXML file name without extension.

replicate

Character with length one, representing the analysis replicate group name.

blank

Character with length one, representing the associated blank replicate group name.

file

mzML or mzXML full file path (with extension).

type

Character with length one. Possible values as "MS" for only MS1 spectral data, "MS/MS" for tandem spectral data (i.e., MS1 and MS2) or "SRM" for selected reaction monitoring data (i.e., no spectra only chromatograms).

instrument

List with metadata from the instrument used to acquire the data (content is highly vendor dependent).

time_stamp

Character with length one, representing the start time and date of the data acquisition.

spectra_number

Integer with the number of spectra in the file.

spectra_mode

Character with length one. Possible values are centroid , profile or NA for centroided data, data in profile mode or mode not defined or absence spectra, respectively.

spectra_levels

Integer vector with the number of levels in the spectra (i.e., 1 or 1 and 2).

mz_low

Numeric length one with the lowest m/z value detected in the spectra, in Da.

mz_high

Numeric length one with the highest m/z value detected in the spectra, in Da.

rt_start

Numeric length one with the run start time, in seconds.

rt_end

Numeric length one with the run end time, in seconds.

polarity

Character length one. Possible values are positive, negative or both (the latter refers to polarity switching acquisition).

chromatograms_number

Integer with the number of chromatograms in the file.

ion_mobility

Logical length one for presence or absence of drift separation from ion mobility.

tic

data.table with the total ion chromatogram.

bpc

data.table with the base peak chromatogram.

spectra

data.table with the raw spectra (only present if loaded).

chromatograms

data.table with the raw chromatograms (only present if loaded).

features

data.table with the features from data processing.

metadata

List with flexible storage for experimental metadata (e.g., concentration, location, etc.).

msAnalysis

A msAnalysis S3 class object.

value

A list to be checked and/or converted to msAnalysis S3 class.

files

A character vector with mzML or mzXML full file path/s. Alternatively, a data.frame with the column/s file, replicate and blank with the full file path/s, the replicate group name/s (string) and the associated blank replicate group name (string).

runParallel

Logical length one. Set to TRUE for parallel processing. Note that parallel processing is not always faster due to overhead.

Value

An msAnalysis S3 class object.

Methods (by generic)

validate(msAnalysis): Validates an msAnalysis S3 class object, returning a logical value of length one.

Functions

as.msAnalysis(): Converts the argument in an msAnalysis S3 class object.

parse.msAnalysis(): Parses information from mzML or mzXML file/s and returns a list with msAnalysis S3 class object/s. On error, returns NULL.