msData R6 class and methods

The msData R6 class is a framework with methods for parsing, processing, visualizing and storing MS data.

References

Helmus R, ter Laak TL, van Wezel AP, de Voogt P, Schymanski EL (2021). “patRoon: open source software platform for environmental mass spectrometry based non-target screening.” Journal of Cheminformatics, 13(1). doi:10.1186/s13321-020-00477-w .

Methods

Public methods

msData$new()

msData$print()

msData$get_headers()

msData$get_analyses()

msData$get_number_analyses()

msData$get_overview()

msData$get_analysis_names()

msData$get_replicate_names()

msData$get_blank_names()

msData$get_polarities()

msData$get_file_paths()

msData$get_mz_low()

msData$get_mz_high()

msData$get_rt_start()

msData$get_rt_end()

msData$get_spectra_mode()

msData$get_spectra_levels()

msData$get_spectra()

msData$get_chromatograms()

msData$get_tic()

msData$get_bpc()

msData$get_eic()

msData$get_ms1()

msData$get_ms2()

msData$get_settings()

msData$get_features()

msData$get_features_eic()

msData$get_features_ms1()

msData$get_features_ms2()

msData$get_alignment()

msData$get_groups()

msData$get_groups_ms1()

msData$get_groups_ms2()

msData$add_headers()

msData$add_settings()

msData$add_analyses()

msData$add_replicate_names()

msData$add_blank_names()

msData$add_features()

msData$add_groups()

msData$add_alignment()

msData$load_spectra()

msData$load_chromatograms()

msData$load_features_ms1()

msData$load_features_ms2()

msData$load_groups_ms1()

msData$load_groups_ms2()

msData$remove_headers()

msData$remove_settings()

msData$remove_analyses()

msData$remove_features()

msData$remove_features_ms1()

msData$remove_features_ms2()

msData$remove_groups()

msData$remove_groups_ms1()

msData$remove_groups_ms2()

msData$remove_alignment()

msData$subset_analyses()

msData$subset_features()

msData$subset_groups()

msData$has_analyses()

msData$has_loaded_spectra()

msData$has_loaded_chromatograms()

msData$has_loaded_features_ms1()

msData$has_loaded_features_ms2()

msData$has_loaded_groups_ms1()

msData$has_loaded_groups_ms2()

msData$has_settings()

msData$has_features()

msData$has_alignment()

msData$has_groups()

msData$plot_spectra()

msData$plot_xic()

msData$plot_eic()

msData$plot_tic()

msData$plot_bpc()

msData$plot_ms2()

msData$plot_ms1()

msData$plot_features()

msData$map_features()

msData$plot_features_ms1()

msData$plot_features_ms2()

msData$plot_alignment()

msData$plot_groups()

msData$plot_groups_ms1()

msData$plot_groups_ms2()

msData$plot_groups_overview()

msData$find_features()

msData$group_features()

msData$as_features_patRoon()

msData$check_analyses_argument()

msData$check_correspondence()

msData$save_headers()

msData$save_settings()

msData$save_analyses()

msData$save_groups()

msData$save()

msData$import_headers()

msData$import_settings()

msData$import_analyses()

msData$import_groups()

msData$processing_function_calls()

msData$clone()

Method new()

Creates an msData class object. When headers are not given (i.e., NULL), a default headers S3 class object is generated with name as NA_character, path as get_wd() and date as Sys.time(). See ?headers for more information.

Usage

msData$new(
  files = NULL,
  runParallel = FALSE,
  headers = NULL,
  settings = NULL,
  analyses = NULL,
  groups = NULL,
  alignment = NULL
)

Arguments

files

A character vector with full file path/s of mzML or mzXML file/s or a data.frame with the columns file, replicate and blank with the full file path of mzML or mzXML file/s (character), the replicate group name (character) and the associated blank replicate group name (character). Note the file paths must have the extension .mzML or .mzXML.

runParallel

Logical length one. Set to TRUE for parallel processing. Note that parallel processing is not always faster due to overhead.

headers

A named list with headers or a headers S3 class object. Each list element must have length one. There is not type restrictions except for name (must be type character), path (must be type character) and date (must be class POSIXct or class POSIXt). See more information in ?headers.

settings

A settings S3 class object or a list with settings S3 class objects as elements (see ?settings for more information).

settings

A settings S3 class object or a list with settings S3 class objects as elements (see ?settings for more information).

analyses

An msAnalysis S3 class object or a list with msAnalysis S3 class objects as elements (see ?msAnalysis for more information).

analyses

An msAnalysis S3 class object or a list with msAnalysis S3 class objects as elements (see ?msAnalysis for more information).

groups

data.table with feature groups from correspondence across MS analyses as obtained by the msData method get_groups(). Note that correspondence of features across MS analyses is performed with the method group_features() from the msData class.

groups

data.table with feature groups from correspondence across MS analyses as obtained by the msData method get_groups(). Note that correspondence of features across MS analyses is performed with the method group_features() from the msData class.

alignment

X.

Returns

A new msData class object.

Method print()

Prints a summary of the msData object in the console.

Usage

msData$print()

Returns

Console text.

Method get_headers()

Method to get the headers of the msData object.

Usage

msData$get_headers(value = NULL)

Arguments

value

A character vector with the name of the entry. Possible values are name, author, description, path and date.

Returns

The headers list as defined by value.

Method get_analyses()

Method to get specific analyses from the msData object.

Usage

msData$get_analyses(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

The list of analyses defined by value.

Method get_number_analyses()

Method to get the number of analysis in the msData object.

Usage

msData$get_number_analyses()

Returns

An integer value.

Method get_overview()

Method to get the overview data.frame with all the analysis types, names, replicates, associated blank replicates, polarities and full file paths.

Usage

msData$get_overview()

Returns

A data.frame with columns type, analysis, replicate, blank polarity and file.

Method get_analysis_names()

Method to get the analysis names.

Usage

msData$get_analysis_names(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

A character vector.

Method get_replicate_names()

Method to get the analysis replicate names.

Usage

msData$get_replicate_names(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

A character vector.

Method get_blank_names()

Method to get the analysis blank replicate names.

Usage

msData$get_blank_names(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

A character vector.

Method get_polarities()

Method to get the polarity of the analyses.

Usage

msData$get_polarities(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

A character vector.

Method get_file_paths()

Method to get the polarity of the analyses.

Usage

msData$get_file_paths(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

A character vector.

Method get_mz_low()

Method to get the polarity of the analyses.

Usage

msData$get_mz_low(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

A character vector.

Method get_mz_high()

Method to get the polarity of the analyses.

Usage

msData$get_mz_high(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

A character vector.

Method get_rt_start()

Method to get the polarity of the analyses.

Usage

msData$get_rt_start(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

A character vector.

Method get_rt_end()

Method to get the polarity of the analyses.

Usage

msData$get_rt_end(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

A character vector.

Method get_spectra_mode()

Method to get the polarity of the analyses.

Usage

msData$get_spectra_mode(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

A character vector.

Method get_spectra_levels()

Method to get the spectra levels of the analyses.

Usage

msData$get_spectra_levels(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

A list for each analysis with an integer vector.

Method get_spectra()

Method to get spectra from the MS analyses.

Usage

msData$get_spectra(
  analyses = NULL,
  levels = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  allTraces = TRUE,
  isolationWindow = 1.3,
  minIntensityMS1 = 0,
  minIntensityMS2 = 0,
  runParallel = FALSE
)

Arguments

analyses

X.

levels

Name.

mz

X.

rt

X.

ppm

X.

sec

X.

id

X.

allTraces

Logical, when TRUE all level 2 data is returned. When FALSE and level has 2, only the MS2 traces of MS1 targets are returned, using the isolationWindow to calculate the mass window.

isolationWindow

X.

minIntensityMS1

X.

minIntensityMS2

X.

runParallel

X.

Returns

A data.frame with spectra for each analyses and targets when defined.

Method get_chromatograms()

Method to get spectra from the MS analyses.

Usage

msData$get_chromatograms(
  analyses = NULL,
  minIntensity = 0,
  runParallel = FALSE
)

Arguments

analyses

X.

minIntensity

X.

runParallel

X.

Returns

A data.frame with spectra.

Method get_tic()

Method to get the total ion chromatograms (TIC) from the analyses.

Usage

msData$get_tic(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

A character vector.

Method get_bpc()

Method to get the base peak chromatograms (BPC) from the analyses.

Usage

msData$get_bpc(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

A character vector.

Method get_eic()

Method to get extract ion chromatograms (EIC) from the analyses based on targets.

Usage

msData$get_eic(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  runParallel = FALSE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

mz

X.

rt

X.

ppm

X.

sec

X.

id

X.

runParallel

X.

Returns

A data.frame.

Method get_ms1()

Method to get MS1 data from the analyses based on targets.

Usage

msData$get_ms1(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  mzClust = 0.003,
  verbose = FALSE,
  minIntensity = 1000,
  runParallel = FALSE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

mz

X.

rt

X.

ppm

X.

sec

X.

id

X.

mzClust

X.

verbose

X.

minIntensity

X.

runParallel

X.

Returns

A data.frame.

Method get_ms2()

Method to get MS2 data from the analyses based on targets.

Usage

msData$get_ms2(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  isolationWindow = 1.3,
  mzClust = 0.005,
  verbose = FALSE,
  minIntensity = 0,
  runParallel = FALSE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

mz

X.

rt

X.

ppm

X.

sec

X.

id

X.

isolationWindow

X.

mzClust

X.

verbose

X.

minIntensity

X.

runParallel

X.

Returns

A data.frame.

Method get_settings()

Method to get settings from analyses.

Usage

msData$get_settings(call = NULL)

Arguments

call

A string with the name of function call.

Returns

A data.frame.

Method get_features()

Method to get features from analyses.

Usage

msData$get_features(
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  filtered = FALSE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

features

X.

mass

X.

mz

X.

rt

X.

ppm

X.

sec

X.

filtered

X.

Returns

A data.frame.

Method get_features_eic()

Method to get EIC of features from analyses.

Usage

msData$get_features_eic(
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  rtExpand = 120,
  mzExpand = 0.005,
  filtered = FALSE,
  runParallel = FALSE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

features

X.

mass

X.

mz

X.

rt

X.

ppm

X.

sec

X.

rtExpand

X.

mzExpand

X.

filtered

X.

runParallel

X.

Returns

A data.frame/data.table.

Method get_features_ms1()

Method to get an averaged MS1 spectrum for features in analyses.

Usage

msData$get_features_ms1(
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  rtWindow = c(-2, 2),
  mzWindow = c(-5, 100),
  mzClust = 0.003,
  minIntensity = 1000,
  verbose = FALSE,
  filtered = FALSE,
  loadedMS1 = TRUE,
  runParallel = FALSE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

features

X.

mass

X.

mz

X.

rt

X.

ppm

X.

sec

X.

rtWindow

X.

mzWindow

X.

mzClust

X.

minIntensity

X.

verbose

X.

filtered

X.

loadedMS1

X.

runParallel

X.

Returns

A data.frame/data.table.

Method get_features_ms2()

Method to get an averaged MS2 spectrum for features in analyses.

Usage

msData$get_features_ms2(
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  isolationWindow = 1.3,
  mzClust = 0.003,
  minIntensity = 0,
  verbose = FALSE,
  filtered = FALSE,
  loadedMS2 = TRUE,
  runParallel = FALSE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

features

X.

mass

X.

mz

X.

rt

X.

ppm

X.

sec

X.

isolationWindow

X.

mzClust

X.

minIntensity

X.

verbose

X.

filtered

X.

loadedMS2

X.

runParallel

X.

Returns

A data.frame/data.table.

Method get_alignment()

Method to get alignment.

Usage

msData$get_alignment()

Returns

A data.frame.

Method get_groups()

Method to get feature groups from analyses.

Usage

msData$get_groups(
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  filtered = FALSE,
  onlyIntensities = FALSE,
  average = FALSE
)

Arguments

groups

X.

mass

X.

mz

X.

rt

X.

ppm

X.

sec

X.

filtered

X.

onlyIntensities

X.

average

X.

Returns

A data.frame/data.table.

Method get_groups_ms1()

Method to get an averaged MS1 spectrum for feature groups in analyses.

Usage

msData$get_groups_ms1(
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  rtWindow = c(-2, 2),
  mzWindow = c(-5, 90),
  mzClustFeatures = 0.003,
  minIntensityFeatures = 1000,
  loadedFeaturesMS1 = TRUE,
  mzClustGroups = 0.003,
  minIntensityGroups = 1000,
  groupBy = "groups",
  verbose = FALSE,
  filtered = FALSE,
  loadedGroupsMS1 = TRUE,
  runParallel = FALSE
)

Arguments

groups

X.

mass

X.

mz

X.

rt

X.

ppm

X.

sec

X.

rtWindow

X.

mzWindow

X.

mzClustFeatures

X.

minIntensityFeatures

X.

loadedFeaturesMS1

X.

mzClustGroups

X.

minIntensityGroups

X.

groupBy

X.

verbose

X.

filtered

X.

loadedGroupsMS1

X.

runParallel

X.

Returns

A data.frame/data.table.

Method get_groups_ms2()

Method to get an averaged MS2 spectrum for feature groups in analyses.

Usage

msData$get_groups_ms2(
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  isolationWindow = 1.3,
  mzClustFeatures = 0.003,
  minIntensityFeatures = 100,
  loadedFeaturesMS2 = TRUE,
  mzClustGroups = 0.003,
  minIntensityGroups = 100,
  groupBy = "groups",
  verbose = FALSE,
  filtered = FALSE,
  loadedGroupsMS2 = TRUE,
  runParallel = FALSE
)

Arguments

groups

X.

mass

X.

mz

X.

rt

X.

ppm

X.

sec

X.

isolationWindow

X.

mzClustFeatures

X.

minIntensityFeatures

X.

loadedFeaturesMS2

X.

mzClustGroups

X.

minIntensityGroups

X.

groupBy

X.

verbose

X.

filtered

X.

loadedGroupsMS2

X.

runParallel

X.

Returns

A data.frame/data.table.

Method add_headers()

Method to add headers information to the msData object. If an argument or element name is given, it must be type character. If an argument or element path is given, it must be type character and exist. If an argument or element date is given, it must be class POSIXct or POSIXt. If given date is character, conversion to class POSIXct or POSIXt is attempted. See ?headers for more information.

Usage

msData$add_headers(...)

Arguments

...

Arguments to be added as headers. Note that all given arguments must be of length one. There is not type limitation. The name of the argument becomes the name of the header entry. Alternatively, a single list argument with the headers (all length one) can be given.

Returns

Invisible.

Method add_settings()

Method to add processing settings to the msData object.

Usage

msData$add_settings(settings = NULL)

Arguments

settings

A settings S3 class object or a list with settings S3 class objects as elements (see ?settings for more information).

settings

A settings S3 class object or a list with settings S3 class objects as elements (see ?settings for more information).

Returns

Invisible.

Method add_analyses()

Method to add analyses to the msData object.

Usage

msData$add_analyses(analyses = NULL)

Arguments

analyses

An msAnalysis S3 class object or a list with msAnalysis S3 class objects as elements (see ?msAnalysis for more information).

analyses

An msAnalysis S3 class object or a list with msAnalysis S3 class objects as elements (see ?msAnalysis for more information).

Returns

Invisible.

Method add_replicate_names()

Method to add or redefine the analysis replicate names. Changes the msData object.

Usage

msData$add_replicate_names(value = NULL)

Arguments

value

A character vector with the analysis replicate names. Must be of the same length as the number of analyses.

Returns

Invisible.

Method add_blank_names()

Method to add or redefine the analysis blank replicate names. Changes the msData object.

Usage

msData$add_blank_names(value = NULL)

Arguments

value

A character vector with the analysis blank replicate names. Must be of the same length as the number of analyses.

Returns

Invisible.

Method add_features()

Method to add features to each analysis in the msData object.

Usage

msData$add_features(features = NULL)

Arguments

features

X.

Returns

Invisible.

Method add_groups()

Method to add feature groups in the msData object.

Usage

msData$add_groups(groups = NULL)

Arguments

groups

data.table with feature groups from correspondence across MS analyses as obtained by the msData method get_groups(). Note that correspondence of features across MS analyses is performed with the method group_features() from the msData class.

groups

data.table with feature groups from correspondence across MS analyses as obtained by the msData method get_groups(). Note that correspondence of features across MS analyses is performed with the method group_features() from the msData class.

Returns

Invisible.

Method add_alignment()

Method to add time alignment results in the msData object.

Usage

msData$add_alignment(alignment = NULL)

Arguments

alignment

list.

Returns

Invisible.

Method load_spectra()

Method to load all spectra from analyses to the msData object.

Usage

msData$load_spectra(runParallel = FALSE)

Arguments

runParallel

X.

Returns

Invisible.

Method load_chromatograms()

Method to load all chromatograms from analyses to the msData object.

Usage

msData$load_chromatograms(runParallel = FALSE)

Arguments

runParallel

X.

Returns

Invisible.

Method load_features_ms1()

Method to load and average MS1 spectra from features in the analyses.

Usage

msData$load_features_ms1(settings = NULL)

Arguments

settings

X.

Returns

Invisible.

Method load_features_ms2()

Method to load and average MS2 spectra from features in the analyses.

Usage

msData$load_features_ms2(settings = NULL)

Arguments

settings

X.

Returns

Invisible.

Method load_groups_ms1()

Method to load and average MS1 spectra from feature groups. If MS1 spectra from features are already loaded, the feature MS1 spectra is used for averaging into the respective feature group. If features MS1 are not present, settings for loading and averaging features MS1 spectra (i.e., settings with call name "load_features_ms1") must be given in the settingsFeatures argument or added beforehand to the msData object.

Usage

msData$load_groups_ms1(settings = NULL, settingsFeatures = NULL)

Arguments

settings

X.

settingsFeatures

X.

Returns

Invisible.

Method load_groups_ms2()

Method to load and average MS2 spectra from feature groups.

Usage

msData$load_groups_ms2(settings = NULL, settingsFeatures = NULL)

Arguments

settings

X.

settingsFeatures

X.

Returns

Invisible.

Method remove_headers()

Removes headers entries from the msData object. Note that the name, path and date headers cannot be removed.

Usage

msData$remove_headers(value = NULL)

Arguments

value

A character vector with the names of the headers entries to be removed.

Returns

Invisible.

Method remove_settings()

Removes settings from the msData object.

Usage

msData$remove_settings(call = NULL)

Arguments

call

A character vector with the settings call name. When call is NULL all settings are removed.

Returns

Invisible.

Method remove_analyses()

Removes analyses from the msData object. Note that unique feature groups from the removed analyses are also removed.

Usage

msData$remove_analyses(analyses = NULL)

Arguments

analyses

A character vector with the names or indices of the analyses to be removed.

Returns

Invisible.

Method remove_features()

Remove feature from the msData object.

Usage

msData$remove_features(features = NULL, filtered = FALSE)

Arguments

features

A data.frame with columns analysis and feature representing the analysis name and the id of the features to remove, respectively.

filtered

X.

Returns

Invisible.

Method remove_features_ms1()

Removes loaded MS1 spectra from features in the analyses of the msData object. In practice, the column ms1 in the features data.table of each analysis object is removed.

Usage

msData$remove_features_ms1()

Returns

Invisible.

Method remove_features_ms2()

Removes loaded MS2 spectra from features in the analyses of the msData object. In practice, the column ms2 in the features data.table of each analysis object is removed.

Usage

msData$remove_features_ms2()

Returns

Invisible.

Method remove_groups()

Remove feature groups from the msData object.

Usage

msData$remove_groups(groups = NULL, filtered = FALSE)

Arguments

groups

X.

filtered

X.

Returns

Invisible.

Method remove_groups_ms1()

Removes loaded MS1 spectra from feature groups in the msData object. In practice, the column ms1 in the groups data.table is removed.

Usage

msData$remove_groups_ms1()

Returns

Invisible.

Method remove_groups_ms2()

Removes loaded MS2 spectra from feature groups in the msData object. In practice, the column ms2 in the groups data.table is removed.

Usage

msData$remove_groups_ms2()

Returns

Invisible.

Method remove_alignment()

Removes alignment results from the msData object.

Usage

msData$remove_alignment()

Returns

Invisible.

Method subset_analyses()

Subsets an msData object on analyses.

Usage

msData$subset_analyses(analyses = NULL)

Arguments

analyses

X.

Returns

A new cloned msData object with only the analyses as defined by the analyses argument.

Method subset_features()

Subsets an msData object on features from analyses.

Usage

msData$subset_features(features = NULL)

Arguments

features

A data.frame with columns analysis and feature representing the analysis name and the id of the features to keep in the new msData object, respectively.

Returns

A new cloned msData object with only the features as defined by the features argument.

Method subset_groups()

Subsets an msData object on groups from correspondence of features across analyses. Note that when sub-setting groups, features that lose correspondence are not removed but filtered with "grouping" added as filter category/tag. Filtered features can be removed with the method msData$remove_features(filtered = TRUE).

Usage

msData$subset_groups(groups = NULL)

Arguments

groups

X.

Returns

A new cloned msData object with only the groups as defined by the groups argument.

Method has_analyses()

Method to check of the msData object has analyses.

Usage

msData$has_analyses()

Returns

Logical value.

Method has_loaded_spectra()

Method to check for loaded spectra in given analyses names/indices.

Usage

msData$has_loaded_spectra(analyses = NULL)

Arguments

analyses

The analyses names/indices to check for loaded spectra.

Returns

Logical value.

Method has_loaded_chromatograms()

Method to check for loaded chromatograms in given analyses names/indices.

Usage

msData$has_loaded_chromatograms(analyses = NULL)

Arguments

analyses

The analyses names/indices to check for loaded chromatograms.

Returns

Logical value.

Method has_loaded_features_ms1()

Method to check for loaded features MS1 in given analyses names/indices.

Usage

msData$has_loaded_features_ms1(analyses = NULL)

Arguments

analyses

The analyses names/indices to check for loaded chromatograms.

Returns

Logical value.

Method has_loaded_features_ms2()

Method to check for loaded features MS2 in given analyses names/indices.

Usage

msData$has_loaded_features_ms2(analyses = NULL)

Arguments

analyses

The analyses names/indices to check for loaded chromatograms.

Returns

Logical value.

Method has_loaded_groups_ms1()

Method to check for loaded feature groups MS1.

Usage

msData$has_loaded_groups_ms1()

Returns

Logical value.

Method has_loaded_groups_ms2()

Method to check for loaded feature groups MS2.

Usage

msData$has_loaded_groups_ms2()

Returns

Logical value.

Method has_settings()

Method to check if there are processing settings in the msData object.

Usage

msData$has_settings(call = NULL)

Arguments

call

A string with the name of function call.

Returns

Logical value.

Method has_features()

Method to check if given analyses have features.

Usage

msData$has_features(analyses = NULL)

Arguments

analyses

The analyses names/indices to check for loaded spectra.

Returns

Logical value.

Method has_alignment()

Method to check if there is alignment of retention time from grouping features across analyses.

Usage

msData$has_alignment()

Returns

Logical value.

Method has_groups()

Method to check if there are feature groups from grouping features across analyses.

Usage

msData$has_groups()

Returns

Logical value.

Method plot_spectra()

Plots spectra for given MS analyses.

Usage

msData$plot_spectra(
  analyses = NULL,
  levels = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  allTraces = TRUE,
  isolationWindow = 1.3,
  minIntensityMS1 = 0,
  minIntensityMS2 = 0,
  runParallel = FALSE,
  colorBy = "analyses"
)

Arguments

analyses

X.

levels

Name.

mz

X.

rt

X.

ppm

X.

sec

X.

id

X.

allTraces

Logical, when TRUE all level 2 data is returned. When FALSE and level has 2, only the MS2 traces of MS1 targets are returned, using the preMZ value and the isolationWindow.

isolationWindow

X.

minIntensityMS1

X.

minIntensityMS2

X.

runParallel

X.

colorBy

X.

Returns

A 3D interactive plot.

Method plot_xic()

Method to plot extract ion chromatograms (EIC) and m/z vs retention time from the analyses.

Usage

msData$plot_xic(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  runParallel = FALSE,
  legendNames = NULL,
  plotTargetMark = TRUE,
  targetsMark = NULL,
  ppmMark = 5,
  secMark = 10,
  numberRows = 1
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

mz

X.

rt

X.

ppm

X.

sec

X.

id

X.

runParallel

X.

legendNames

x.

plotTargetMark

x.

targetsMark

x.

ppmMark

x.

secMark

x.

numberRows

x.

Returns

A data.frame.

Method plot_eic()

Method to plot extract ion chromatograms (EIC) from the analyses.

Usage

msData$plot_eic(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

mz

X.

rt

X.

ppm

X.

sec

X.

id

X.

runParallel

X.

legendNames

x.

title

x.

colorBy

x.

interactive

x.

Returns

A plot.

Method plot_tic()

Method to plot total ion chromatograms (TIC) of the analyses.

Usage

msData$plot_tic(
  analyses = NULL,
  title = NULL,
  colorBy = "analyses",
  interactive = TRUE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

title

x.

colorBy

x.

interactive

x.

Returns

A plot.

Method plot_bpc()

Method to plot base peak chromatograms (BPC) of the analyses.

Usage

msData$plot_bpc(
  analyses = NULL,
  title = NULL,
  colorBy = "analyses",
  interactive = TRUE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

title

x.

colorBy

x.

interactive

x.

Returns

A plot.

Method plot_ms2()

Method to plot MS2 spectra from the analyses based on targets.

Usage

msData$plot_ms2(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  isolationWindow = 1.3,
  mzClust = 0.005,
  verbose = FALSE,
  minIntensity = 0,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

mz

X.

rt

X.

ppm

X.

sec

X.

id

X.

isolationWindow

X.

mzClust

X.

verbose

X.

minIntensity

X.

runParallel

X.

legendNames

x.

title

x.

colorBy

x.

interactive

x.

Returns

A plot.

Method plot_ms1()

Method to plot MS1 spectra from the analyses based on targets.

Usage

msData$plot_ms1(
  analyses = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  id = NULL,
  mzClust = 0.001,
  verbose = FALSE,
  minIntensity = 1000,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

mz

X.

rt

X.

ppm

X.

sec

X.

id

X.

mzClust

X.

verbose

X.

minIntensity

X.

runParallel

X.

legendNames

x.

title

x.

colorBy

x.

interactive

x.

Returns

A plot.

Method plot_features()

Method to plot features from analyses.

Usage

msData$plot_features(
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  rtExpand = 120,
  mzExpand = 0.005,
  filtered = FALSE,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

features

X.

mass

X.

mz

X.

rt

X.

ppm

X.

sec

X.

rtExpand

X.

mzExpand

X.

filtered

X.

runParallel

X.

legendNames

x.

title

x.

colorBy

x.

interactive

x.

Returns

plot.

Method map_features()

Method to map retention time and m/z of features from analyses.

Usage

msData$map_features(
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  filtered = FALSE,
  xlim = 30,
  ylim = 0.05,
  showLegend = TRUE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

features

X.

mass

X.

mz

X.

rt

X.

ppm

X.

sec

X.

filtered

X.

xlim

X.

ylim

X.

showLegend

X.

legendNames

x.

title

x.

colorBy

x.

interactive

x.

Returns

A plot.

Method plot_features_ms1()

Method to plot MS1 spectra from features in the analyses.

Usage

msData$plot_features_ms1(
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  rtWindow = c(-2, 2),
  mzWindow = c(-5, 100),
  mzClust = 0.003,
  minIntensity = 1000,
  verbose = FALSE,
  filtered = FALSE,
  loadedMS1 = TRUE,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

features

X.

mass

X.

mz

X.

rt

X.

ppm

X.

sec

X.

rtWindow

X.

mzWindow

X.

mzClust

X.

minIntensity

X.

verbose

X.

filtered

X.

loadedMS1

X.

runParallel

X.

legendNames

x.

title

x.

colorBy

x.

interactive

x.

Returns

A plot.

Method plot_features_ms2()

Method to plot MS2 spectra from features in the analyses.

Usage

msData$plot_features_ms2(
  analyses = NULL,
  features = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  isolationWindow = 1.3,
  mzClust = 0.005,
  minIntensity = 0,
  verbose = FALSE,
  filtered = FALSE,
  loadedMS2 = TRUE,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

features

X.

mass

X.

mz

X.

rt

X.

ppm

X.

sec

X.

isolationWindow

X.

mzClust

X.

minIntensity

X.

verbose

X.

filtered

X.

loadedMS2

X.

runParallel

X.

legendNames

x.

title

x.

colorBy

x.

interactive

x.

Returns

A plot.

Method plot_alignment()

Plots the results from the retention time alignment across analyses.

Usage

msData$plot_alignment()

Returns

A plot with the retention time alignment differences for each sample.

Method plot_groups()

Method to plot feature groups EIC.

Usage

msData$plot_groups(
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  rtExpand = 120,
  mzExpand = 0.005,
  filtered = FALSE,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)

Arguments

groups

X.

mass

X.

mz

X.

rt

X.

ppm

X.

sec

X.

rtExpand

X.

mzExpand

X.

filtered

X.

runParallel

X.

legendNames

x.

title

x.

colorBy

x.

interactive

x.

Returns

A plot.

Method plot_groups_ms1()

Method to plot MS1 spectra from features in the analyses.

Usage

msData$plot_groups_ms1(
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  rtWindow = c(-2, 2),
  mzWindow = c(-5, 90),
  mzClustFeatures = 0.005,
  minIntensityFeatures = 1000,
  loadedFeaturesMS1 = TRUE,
  mzClustGroups = 0.005,
  minIntensityGroups = 1000,
  verbose = FALSE,
  filtered = FALSE,
  loadedGroupsMS1 = TRUE,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)

Arguments

groups

X.

mass

X.

mz

X.

rt

X.

ppm

X.

sec

X.

rtWindow

X.

mzWindow

X.

mzClustFeatures

X.

minIntensityFeatures

X.

loadedFeaturesMS1

X.

mzClustGroups

X.

minIntensityGroups

X.

verbose

X.

filtered

X.

loadedGroupsMS1

X.

runParallel

X.

legendNames

x.

title

x.

colorBy

x.

interactive

x.

groupBy

X.

Returns

A plot.

Method plot_groups_ms2()

Method to plot MS1 spectra from features in the analyses.

Usage

msData$plot_groups_ms2(
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  isolationWindow = 1.3,
  mzClustFeatures = 0.003,
  minIntensityFeatures = 100,
  loadedFeaturesMS2 = TRUE,
  mzClustGroups = 0.003,
  minIntensityGroups = 100,
  verbose = FALSE,
  filtered = FALSE,
  loadedGroupsMS2 = TRUE,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  colorBy = "targets",
  interactive = TRUE
)

Arguments

groups

X.

mass

X.

mz

X.

rt

X.

ppm

X.

sec

X.

isolationWindow

X.

mzClustFeatures

X.

minIntensityFeatures

X.

loadedFeaturesMS2

X.

mzClustGroups

X.

minIntensityGroups

X.

verbose

X.

filtered

X.

loadedGroupsMS2

X.

runParallel

X.

legendNames

x.

title

x.

colorBy

x.

interactive

x.

groupBy

X.

Returns

A plot.

Method plot_groups_overview()

Method to give an overview of the EIC, alignment and intensity variance from features within target feature groups.

Usage

msData$plot_groups_overview(
  analyses = NULL,
  groups = NULL,
  mass = NULL,
  mz = NULL,
  rt = NULL,
  ppm = 20,
  sec = 60,
  rtExpand = 120,
  mzExpand = 0.005,
  filtered = FALSE,
  runParallel = FALSE,
  legendNames = NULL,
  title = NULL,
  heights = c(0.35, 0.5, 0.15)
)

Arguments

analyses

X.

groups

X.

mass

X.

mz

X.

rt

X.

ppm

X.

sec

X.

rtExpand

X.

mzExpand

X.

filtered

X.

runParallel

X.

legendNames

x.

title

x.

heights

x.

Returns

A plot.

Method find_features()

Finds features (i.e., chromatographic peaks) from MS data in an msData class object. The function uses the patRoon package for peak finding, enabling the use of several algorithms (see details). Note that the settings call name must be "find_features".

Usage

msData$find_features(settings = NULL)

Arguments

settings

A list object with call name, algorithm and parameters. When not given, settings will be searched within the msData object.

Details

See the findFeatures function from the patRoon package or the reference guide for more information. The following algorithms are available via patRoon: "xcms3", "xcms", "openms", "envipick", "sirius", "kpic2", "safd". The algorithm in the settings should be one of the described. The parameters are given as a list and should match with algorithm requirements. Certain algorithms also require defined MS file formats and data in profile mode.

Returns

Invisible.

Method group_features()

Groups and aligns features across analyses in the msData object. The function uses the patRoon package for grouping features, enabling the use of several algorithms (see details). Note that the settings call name must be "group_features".

Usage

msData$group_features(settings = NULL)

Arguments

settings

A list object with call name, algorithm and parameters. When not given, settings will be searched within the msData object.

Details

See the groupFeatures function from the patRoon package or the reference guide for more information. The following algorithms are possible: "xcms3", "xcms", "openms" or "kpic2". The algorithm slot in the settings should be one of the described. The parameters are given as a list and should match with algorithm requirements.

Returns

Invisible.

Method as_features_patRoon()

Creates an object with S4 class features from the package patRoon with the features in the analyses of the msData object.

Usage

msData$as_features_patRoon()

Returns

An object with S4 class features.

Method check_analyses_argument()

Checks the analyses argument as a character/integer vector to match analyses names or indices from the msData object.

Usage

msData$check_analyses_argument(analyses = NULL)

Arguments

analyses

X.

Returns

A valid character vector with analyses names of NULL.

Method check_correspondence()

Checks the correspondence of features within feature groups in the msData object.

Usage

msData$check_correspondence()

Returns

TRUE or FALSE.

Method save_headers()

Method to save the headers list.

Usage

msData$save_headers(format = "json", name = "headers", path = getwd())

Arguments

format

X.

name

X.

path

X.

Returns

Saves the headers list as the defined format in path and returns invisible.

Method save_settings()

Method to save settings list.

Usage

msData$save_settings(
  call = NULL,
  format = "json",
  name = "settings",
  path = getwd()
)

Arguments

call

x:

format

X.

name

X.

path

X.

Returns

Saves the settings list as the defined format in path and returns invisible.

Method save_analyses()

Method to save the list of analyses.

Usage

msData$save_analyses(
  analyses = NULL,
  format = "json",
  name = "analyses",
  path = getwd()
)

Arguments

analyses

X.

format

X.

name

X.

path

X.

Returns

Saves the list of analyses as the defined format in path and returns invisible.

Method save_groups()

Method to save the feature groups data.table.

Usage

msData$save_groups(format = "json", name = "groups", path = getwd())

Arguments

format

X.

name

X.

path

X.

Returns

Saves the groups data.table as the defined format in the path and returns invisible.

Method save()

Method to save the private fields (i.e., headers, settings, analyses, groups and alignment) of the msData object.

Usage

msData$save(format = "json", name = "msData", path = getwd())

Arguments

format

X.

name

X.

path

X.

Returns

Saves the private fields of the msdata as the defined format in the path and returns invisible.

Method import_headers()

Method to import headers to the msData object from a rds or json file.

Usage

msData$import_headers(file = NA_character_, list = NULL)

Arguments

file

X.

list

X.

Returns

Invisible.

Method import_settings()

Method to import processing settings to the msData object from a rds or json file.

Usage

msData$import_settings(file = NA_character_)

Arguments

file

X.

Returns

Invisible.

Method import_analyses()

Method to import analyses to the msData object from a rds or json file.

Usage

msData$import_analyses(file = NA_character_)

Arguments

file

X.

Returns

Invisible.

Method import_groups()

Method to import feature groups to the msData object from a rds or json file.

Usage

msData$import_groups(file = NA_character_)

Arguments

file

X.

Returns

Invisible.

Method processing_function_calls()

Possible processing function calls.

Usage

msData$processing_function_calls()

Returns

A character vector with ordered possible function calls for data pre and post-processing.

Method clone()

The objects of this class are cloneable with this method.

Usage

msData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.