plot_xic_interactive

Plot traces or profile data for targets using expected m/z and retention time pairs, including deviations.

Usage

plot_xic_interactive(
  xic,
  legendNames = NULL,
  plotTargetMark = TRUE,
  targetsMark = NULL,
  ppmMark = 5,
  secMark = 10,
  numberRows = 1
)

Arguments

xic

A data.table with at least columns: mz, rt and intensity.

legendNames

A character vector with the same length as the unique ids in the column id of the data.table xic.

plotTargetMark

Logical, set to TRUE (the default) to plot a target mark.

targetsMark

A data.frame with columns "mz" and "rt", defining the m/z and retention time values of each targets (i.e., each unique id in the xic table). Note that the number of rows should match with the number of unique target ids.

ppmMark

A numeric vector of length one to define the mass deviation, in ppm, of the target mark. The default is 5 ppm.

secMark

A numeric vector of length one to define the time deviation, in seconds, of the target mark. The default is 10 ppm.

numberRows

A numeric vector of length one to define the number of rows to grid the plots. Note that each target is always plotted in one row for all selected analyses.

Value

An iterative plot of the traces for the requested m/z and retention time pairs.