CoreEngine R6 class and methods

The CoreEngine R6 class is a basic data processor with generic methods for handling data.

Note

If there are ProcessingSettings objects with the same method names it is better to use run_workflow().

The engine data is saved in an sqlite file and loaded in the app. If save file is defined in the engine it is used, otherwise the save file name is automatically set to the engine class name and the date. Changes made in the app can be saved in the sqlite file and then loaded for scripting.

Active bindings

headers

ProjectHeaders S7 class object.

workflow

Workflow S7 class object.

analyses

Analyses S7 class object.

results

List of results.

history

Audit trail of changes.

file

An EngineSaveFile S7 class object. When setting the value it can be also a character with an sqlite file path to save the engine.

Methods

Public methods

CoreEngine$new()

CoreEngine$print()

CoreEngine$print_headers()

CoreEngine$print_analyses()

CoreEngine$print_workflow()

CoreEngine$save()

CoreEngine$load()

CoreEngine$add_analyses()

CoreEngine$remove_analyses()

CoreEngine$has_settings()

CoreEngine$has_analyses()

CoreEngine$has_results()

CoreEngine$run()

CoreEngine$run_workflow()

CoreEngine$export_headers()

CoreEngine$export_workflow()

CoreEngine$export_analyses()

CoreEngine$export()

CoreEngine$import_headers()

CoreEngine$import_workflow()

CoreEngine$import_analyses()

CoreEngine$import()

CoreEngine$run_app()

CoreEngine$clone()

Method new()

Creates a CoreEngine R6 class object.

Usage

CoreEngine$new(file = NULL, headers = NULL, workflow = NULL, analyses = NULL)

Arguments

file

Character of length one with the full path to the sqlite save file of the engine.

headers

A ProjectHeaders S7 class object.

workflow

A Workflow S7 class object.

analyses

An Analyses S7 class object.

Method print()

Prints a summary to the console.

Usage

CoreEngine$print()

Method print_headers()

Prints the headers.

Usage

CoreEngine$print_headers()

Method print_analyses()

Prints the analyses.

Usage

CoreEngine$print_analyses()

Method print_workflow()

Prints the workflow.

Usage

CoreEngine$print_workflow()

Method save()

Saves the engine data as an sqlite file.

Usage

CoreEngine$save(file = NA_character_)

Arguments

file

A string with the full file path of the sqlite file. If NA (the default) and no save file is defined in the engine, the file name is automatically created with the engine class name and the date in the headers.

Method load()

Loads the engine data from an sqlite file.

Usage

CoreEngine$load(file = NA_character_)

Arguments

file

A string with the full file path of the sqlite file containing the engine data saved.

Method add_analyses()

Adds analyses. Note that when adding new analyses, any existing results are removed.

Usage

CoreEngine$add_analyses(analyses = NULL)

Arguments

analyses

A Analysis S3 class object or a list with Analysis S3 class objects as elements (see ?Analysis for more information) or a character vector with full path to analysis files.

Returns

Invisible.

Method remove_analyses()

Removes analyses.

Usage

CoreEngine$remove_analyses(analyses = NULL)

Arguments

analyses

A string or a vector of strings with the name/s or numeric with indices of the analyses to remove.

Returns

Invisible.

Method has_settings()

Checks if there are processing settings, returning TRUE or FALSE.

Usage

CoreEngine$has_settings()

Method has_analyses()

Checks if analyses are present, returning TRUE or FALSE.

Usage

CoreEngine$has_analyses()

Method has_results()

Checks if results are present, returning TRUE or FALSE.

Usage

CoreEngine$has_results(names = NULL)

Arguments

names

A string or a vector of strings with the name/s of the results data. The actual names depends of the applied algorithms. For instance, when algorithms via patRoon are used, the name of the module data is patRoon.

Method run()

Runs a processing method according to the provided settings.

Usage

CoreEngine$run(settings = NULL)

Arguments

settings

A ProcessingSettings object or a character string with the method name of a previously added ProcessingSettings in the engine.

Method run_workflow()

Runs all processing methods in workflow.

Usage

CoreEngine$run_workflow()

Returns

Invisible.

Method export_headers()

Exports the headers as json (the default) or rds.

Usage

CoreEngine$export_headers(format = "json", name = "headers", path = getwd())

Arguments

format

A string. Possible values are json and rds for saving as JSON or RDS formats, respectively.

name

String with the name of the file to be saved.

path

String with the directory of the file to be saved.

Method export_workflow()

Exports the workflow as json (the default) or rds.

Usage

CoreEngine$export_workflow(format = "json", name = "settings", path = getwd())

Arguments

format

A string. Possible values are json and rds for saving as JSON or RDS formats, respectively.

name

String with the name of the file to be saved.

path

String with the directory of the file to be saved.

Method export_analyses()

Exports the analyses as json (the default) or rds.

Usage

CoreEngine$export_analyses(format = "json", name = "analyses", path = getwd())

Arguments

format

A string. Possible values are json and rds for saving as JSON or RDS formats, respectively.

name

String with the name of the file to be saved.

path

String with the directory of the file to be saved.

Method export()

Exports the engine data as as json (the default) or rds.

Usage

CoreEngine$export(format = "json", name = "EngineData", path = getwd())

Arguments

format

A string. Possible values are json and rds for saving as JSON or RDS formats, respectively.

name

String with the name of the file to be saved.

path

String with the directory of the file to be saved.

Method import_headers()

Imports headers from an rds or json file.

Usage

CoreEngine$import_headers(file = NA_character_)

Arguments

file

A json or rds file.

Returns

Invisible.

Method import_workflow()

Imports workflow from an rds or json file.

Usage

CoreEngine$import_workflow(file = NA_character_)

Arguments

file

A json or rds file.

Returns

Invisible.

Method import_analyses()

Imports analyses from an rds or json file.

Usage

CoreEngine$import_analyses(file = NA_character_)

Arguments

file

A json or rds file.

Returns

Invisible.

Method import()

Imports a CoreEngine saved as json or rds.

Usage

CoreEngine$import(file = NA_character_)

Arguments

file

A json or rds file.

Returns

Invisible.

Method run_app()

Runs a Shiny app to explore and manage the engine.

Usage

CoreEngine$run_app()

Method clone()

The objects of this class are cloneable with this method.

Usage

CoreEngine$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.