RamanEngine R6 class and methods

The RamanEngine R6 class is a framework for parsing, processing, inspecting and storing Raman spectroscopic data.

Super class

StreamFind::CoreEngine -> RamanEngine

Methods

Public methods

RamanEngine$new()

RamanEngine$get_overview()

RamanEngine$get_spectra()

RamanEngine$subset_analyses()

RamanEngine$remove_analyses()

RamanEngine$merge_spectra_time_series()

RamanEngine$average_spectra()

RamanEngine$subtract_blank_spectra()

RamanEngine$correct_spectra_baseline()

RamanEngine$bin_spectra()

RamanEngine$subtract_spectra_section()

RamanEngine$delete_spectra_section()

RamanEngine$smooth_spectra()

RamanEngine$normalize_spectra()

RamanEngine$plot_spectra()

RamanEngine$plot_chromatograms()

RamanEngine$processing_methods()

RamanEngine$clone()

Inherited methods

StreamFind::CoreEngine$add_analyses()

StreamFind::CoreEngine$add_blank_names()

StreamFind::CoreEngine$add_headers()

StreamFind::CoreEngine$add_metadata()

StreamFind::CoreEngine$add_replicate_names()

StreamFind::CoreEngine$add_results()

StreamFind::CoreEngine$add_settings()

StreamFind::CoreEngine$get_analyses()

StreamFind::CoreEngine$get_analysis_names()

StreamFind::CoreEngine$get_blank_names()

StreamFind::CoreEngine$get_files()

StreamFind::CoreEngine$get_formats()

StreamFind::CoreEngine$get_headers()

StreamFind::CoreEngine$get_history()

StreamFind::CoreEngine$get_number_analyses()

StreamFind::CoreEngine$get_replicate_names()

StreamFind::CoreEngine$get_results()

StreamFind::CoreEngine$get_results_names()

StreamFind::CoreEngine$get_settings()

StreamFind::CoreEngine$get_settings_names()

StreamFind::CoreEngine$get_types()

StreamFind::CoreEngine$get_workflow_overview()

StreamFind::CoreEngine$has_analyses()

StreamFind::CoreEngine$has_averaged_spectra()

StreamFind::CoreEngine$has_chromatograms()

StreamFind::CoreEngine$has_chromatograms_peaks()

StreamFind::CoreEngine$has_results()

StreamFind::CoreEngine$has_settings()

StreamFind::CoreEngine$has_spectra()

StreamFind::CoreEngine$has_spectra_peaks()

StreamFind::CoreEngine$import()

StreamFind::CoreEngine$import_analyses()

StreamFind::CoreEngine$import_headers()

StreamFind::CoreEngine$import_settings()

StreamFind::CoreEngine$plot_chromatograms_peaks()

StreamFind::CoreEngine$plot_spectra_peaks()

StreamFind::CoreEngine$print()

StreamFind::CoreEngine$print_analyses()

StreamFind::CoreEngine$print_headers()

StreamFind::CoreEngine$print_results()

StreamFind::CoreEngine$print_workflow()

StreamFind::CoreEngine$remove_headers()

StreamFind::CoreEngine$remove_results()

StreamFind::CoreEngine$remove_settings()

StreamFind::CoreEngine$run_workflow()

StreamFind::CoreEngine$save()

StreamFind::CoreEngine$save_analyses()

StreamFind::CoreEngine$save_headers()

StreamFind::CoreEngine$save_settings()

Method new()

Creates an R6 class RamanEngine. Child of CoreEngine R6 class.

Usage

RamanEngine$new(
  files = NULL,
  headers = NULL,
  settings = NULL,
  analyses = NULL,
  results = NULL,
  runParallel = FALSE
)

Arguments

files

Vector with full paths of .asc files from Raman analyses.

headers

A named list with headers or a ProjectHeaders S3 class object. Each list element must have length one. There is no type restrictions except for name (must be type character length 1), author (must be type character length 1), path (must be type character length 1) and date (must be class POSIXct or class POSIXt length 1). See more information in ?ProjectHeaders.

settings

A named list of ProcessingSettings objects or a single ProcessingSettings object. The list names should match the call name of each ProcessingSettings object. Alternatively, a named list with call name, algorithm and parameters to be transformed and added as ProcessingSettings object.

settings

A ProcessingSettings S3 class object with the call name applicable to the processing method. Alternatively, a named list with call, algorithm and parameters to be transformed and used as ProcessingSettings S3 class object. When not given, settings for the processing method will be searched within the engine object.

analyses

A numeric/character vector with the number/name of the analyses.

results

A named list of objects from processing methods.

runParallel

Logical of length one. Set to TRUE for parallel processing. Note that parallel processing is not always faster due to overhead.

Method get_overview()

Gets an overview data.frame of all the analyses.

Usage

RamanEngine$get_overview()

Method get_spectra()

Gets a data.table with spectra from analyses.

Usage

RamanEngine$get_spectra(
  analyses = NULL,
  rt = NULL,
  shift = NULL,
  minIntensity = 0
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

rt

Numeric of length two with the minimum and maximum retention time, in seconds.

shift

Numeric of length two with the minimum and maximum shift, in cm-1.

minIntensity

Numeric value with the minimum intensity.

Method subset_analyses()

Subsets on analyses.

Usage

RamanEngine$subset_analyses(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

A new cloned engine with only the analyses as defined by the analyses argument.

Method remove_analyses()

Removes analyses.

Usage

RamanEngine$remove_analyses(analyses = NULL)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

Returns

Invisible.

Method merge_spectra_time_series()

Merges spectra for given RamanAnalyses from the same chromatographic separation when using LC-Raman coupling.

Usage

RamanEngine$merge_spectra_time_series(settings)

Arguments

settings

A named list of ProcessingSettings objects or a single ProcessingSettings object. The list names should match the call name of each ProcessingSettings object. Alternatively, a named list with call name, algorithm and parameters to be transformed and added as ProcessingSettings object.

settings

A ProcessingSettings S3 class object with the call name applicable to the processing method. Alternatively, a named list with call, algorithm and parameters to be transformed and used as ProcessingSettings S3 class object. When not given, settings for the processing method will be searched within the engine object.

Returns

Invisible.

Method average_spectra()

Averages spectra based on assigned analysis replicates.

Usage

RamanEngine$average_spectra(settings)

Arguments

settings

A named list of ProcessingSettings objects or a single ProcessingSettings object. The list names should match the call name of each ProcessingSettings object. Alternatively, a named list with call name, algorithm and parameters to be transformed and added as ProcessingSettings object.

settings

A ProcessingSettings S3 class object with the call name applicable to the processing method. Alternatively, a named list with call, algorithm and parameters to be transformed and used as ProcessingSettings S3 class object. When not given, settings for the processing method will be searched within the engine object.

Returns

Invisible.

Method subtract_blank_spectra()

Subtracts spectra from correspondent blank analysis replicates.

Usage

RamanEngine$subtract_blank_spectra(settings)

Arguments

settings

A named list of ProcessingSettings objects or a single ProcessingSettings object. The list names should match the call name of each ProcessingSettings object. Alternatively, a named list with call name, algorithm and parameters to be transformed and added as ProcessingSettings object.

settings

A ProcessingSettings S3 class object with the call name applicable to the processing method. Alternatively, a named list with call, algorithm and parameters to be transformed and used as ProcessingSettings S3 class object. When not given, settings for the processing method will be searched within the engine object.

Returns

Invisible.

Method correct_spectra_baseline()

Corrects the spectra baseline.

Usage

RamanEngine$correct_spectra_baseline(settings)

Arguments

settings

A named list of ProcessingSettings objects or a single ProcessingSettings object. The list names should match the call name of each ProcessingSettings object. Alternatively, a named list with call name, algorithm and parameters to be transformed and added as ProcessingSettings object.

settings

A ProcessingSettings S3 class object with the call name applicable to the processing method. Alternatively, a named list with call, algorithm and parameters to be transformed and used as ProcessingSettings S3 class object. When not given, settings for the processing method will be searched within the engine object.

Returns

Invisible.

Method bin_spectra()

Bins spectra in units, according to a given window size.

Usage

RamanEngine$bin_spectra(settings)

Arguments

settings

A named list of ProcessingSettings objects or a single ProcessingSettings object. The list names should match the call name of each ProcessingSettings object. Alternatively, a named list with call name, algorithm and parameters to be transformed and added as ProcessingSettings object.

settings

A ProcessingSettings S3 class object with the call name applicable to the processing method. Alternatively, a named list with call, algorithm and parameters to be transformed and used as ProcessingSettings S3 class object. When not given, settings for the processing method will be searched within the engine object.

Returns

Invisible.

Method subtract_spectra_section()

Subtracts each spectra by a spectra section in each analysis.

Usage

RamanEngine$subtract_spectra_section(settings)

Arguments

settings

A named list of ProcessingSettings objects or a single ProcessingSettings object. The list names should match the call name of each ProcessingSettings object. Alternatively, a named list with call name, algorithm and parameters to be transformed and added as ProcessingSettings object.

settings

A ProcessingSettings S3 class object with the call name applicable to the processing method. Alternatively, a named list with call, algorithm and parameters to be transformed and used as ProcessingSettings S3 class object. When not given, settings for the processing method will be searched within the engine object.

Returns

Invisible.

Method delete_spectra_section()

Deletes a section of spectra in each analysis.

Usage

RamanEngine$delete_spectra_section(settings)

Arguments

settings

A named list of ProcessingSettings objects or a single ProcessingSettings object. The list names should match the call name of each ProcessingSettings object. Alternatively, a named list with call name, algorithm and parameters to be transformed and added as ProcessingSettings object.

settings

A ProcessingSettings S3 class object with the call name applicable to the processing method. Alternatively, a named list with call, algorithm and parameters to be transformed and used as ProcessingSettings S3 class object. When not given, settings for the processing method will be searched within the engine object.

Returns

Invisible.

Method smooth_spectra()

Smooths the spectra in each analysis/replicate.

Usage

RamanEngine$smooth_spectra(settings)

Arguments

settings

A named list of ProcessingSettings objects or a single ProcessingSettings object. The list names should match the call name of each ProcessingSettings object. Alternatively, a named list with call name, algorithm and parameters to be transformed and added as ProcessingSettings object.

settings

A ProcessingSettings S3 class object with the call name applicable to the processing method. Alternatively, a named list with call, algorithm and parameters to be transformed and used as ProcessingSettings S3 class object. When not given, settings for the processing method will be searched within the engine object.

Returns

Invisible.

Method normalize_spectra()

Normalizes spectra in each analysis/replicate.

Usage

RamanEngine$normalize_spectra(settings)

Arguments

settings

A named list of ProcessingSettings objects or a single ProcessingSettings object. The list names should match the call name of each ProcessingSettings object. Alternatively, a named list with call name, algorithm and parameters to be transformed and added as ProcessingSettings object.

settings

A ProcessingSettings S3 class object with the call name applicable to the processing method. Alternatively, a named list with call, algorithm and parameters to be transformed and used as ProcessingSettings S3 class object. When not given, settings for the processing method will be searched within the engine object.

Returns

Invisible.

Method plot_spectra()

Plots spectra for given RamanAnalyses.

Usage

RamanEngine$plot_spectra(
  analyses = NULL,
  rt = NULL,
  shift = NULL,
  minIntensity = 0,
  xVal = "shift",
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  cex = 0.6,
  showLegend = TRUE,
  colorBy = "analyses",
  interactive = TRUE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

rt

Numeric of length two with the minimum and maximum retention time, in seconds.

shift

Numeric of length two with the minimum and maximum shift, in cm-1.

minIntensity

Numeric value with the minimum intensity.

xVal

Character of length one. Possible are "rt" or "shift" for using the retention time or the shift as x axis, respectively.

xLab

A string with the title for the x axis.

yLab

A string with the title for the y axis.

title

A string with the title.

cex

A length one numeric value defining the scale of the legend for non interactive plotting.

showLegend

Logical. Set to TRUE to show legend.

colorBy

A string defining how to legend the plot. Possible values are analyses, targets (the default) or replicates. Other options are possible for specific plotting methods.

interactive

Logical. When TRUE, the data is plotted interactively using plotly.

Returns

A plot.

Method plot_chromatograms()

Plots chromatograms from analyses with spectra coupled to LC.

Usage

RamanEngine$plot_chromatograms(
  analyses = NULL,
  rt = NULL,
  shift = NULL,
  minIntensity = 0,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  cex = 0.6,
  showLegend = TRUE,
  colorBy = "analyses",
  interactive = TRUE
)

Arguments

analyses

A numeric/character vector with the number/name of the analyses.

rt

Numeric of length two with the minimum and maximum retention time, in seconds.

shift

Numeric of length two with the minimum and maximum shift, in cm-1.

minIntensity

Numeric value with the minimum intensity.

xLab

A string with the title for the x axis.

yLab

A string with the title for the y axis.

title

A string with the title.

cex

A length one numeric value defining the scale of the legend for non interactive plotting.

showLegend

Logical. Set to TRUE to show legend.

colorBy

A string defining how to legend the plot. Possible values are analyses, targets (the default) or replicates. Other options are possible for specific plotting methods.

interactive

Logical. When TRUE, the data is plotted interactively using plotly.

Returns

A plot.

Method processing_methods()

A data.table with available data processing methods.

Usage

RamanEngine$processing_methods()

Method clone()

The objects of this class are cloneable with this method.

Usage

RamanEngine$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.