Statistical Analyses (Tabular Data Input)

The StatisticAnalyses class is used for statistical analysis of tabular data.

Usage

StatisticAnalyses(
  analyses = NULL,
  classes = character(),
  concentrations = numeric(),
  ...
)

Arguments

analyses

A character vector with full file path to .csv file with variable names as first row and analyses names as first column or a data.frame or matrix object.

classes

A character vector with the classes of the analyses.

concentrations

A numeric vector with the concentrations of the analyses.

...

Additional arguments passed to the constructor.

Slots

analyses

A data.frame with the analyses data.

type

A character vector with the type of analyses.

classes

A character vector with the classes of the analyses.

concentrations

A numeric vector with the concentrations of the analyses.

info

(getter) A data.frame with the information of the analyses.

number_variables

(getter) A numeric with the number of variables in the analyses.

has_data

(getter) A logical indicating if the analyses has data.

data

(getter/setter) A list with the data of the analyses.

has_model

(getter) A logical indicating if the analyses has a model.

model

(getter/setter) A list with the model of the analyses.

has_test

(getter) A logical indicating if the analyses has a test.

test

(getter) A list with the test of the analyses.

has_prediction

(getter) A logical indicating if the analyses has a prediction.

prediction

(getter) A list with the prediction of the analyses.

has_quantification

(getter) A logical indicating if the analyses has a quantification.

quantification

(getter/setter) A list with the quantification of the analyses.