Function to convert vendor MS files to mzML or mzXML

Function to convert vendor MS files to .mzML or .mzXML using the command line interface of MSConvert from ProteoWizard.

Usage

convert_ms_files(
  files = NULL,
  outputFormat = "mzML",
  outputPath = NULL,
  optList = NULL
)

Arguments

files

A vector with file full path/s for conversion.

outputFormat

A character string with the format of the output file. Possible values are "mzML" and "mzXML".

outputPath

The directory to place the output file. When NULL the directory of the original file is used.

optList

A list of further options to be passed to the respective conversion tool. See documentation in https://proteowizard.sourceforge.io/tools/msconvert.html for possible commands.

Note

When giving paths make sure that single backslash is changed to double backslash or to single forward slash.

References

Kessner D, Chambers M, Burke R, Agus D, Mallick P (2008). “ProteoWizard: open source software for rapid proteomics tools development.” Bioinformatics, 24(21), 2534--2536. doi:10.1093/bioinformatics/btn323 , http://www.ncbi.nlm.nih.gov/pubmed/18606607.

Chambers MC, Maclean B, Burke R, Amodei D, Ruderman DL, Neumann S, Gatto L, Fischer B, Pratt B, Egertson J, Hoff K, Kessner D, Tasman N, Shulman N, Frewen B, Baker TA, Brusniak M, Paulse C, Creasy D, Flashner L, Kani K, Moulding C, Seymour SL, Nuwaysir LM, Lefebvre B, Kuhlmann F, Roark J, Rainer P, Detlev S, Hemenway T, Huhmer A, Langridge J, Connolly B, Chadick T, Holly K, Eckels J, Deutsch EW, Moritz RL, Katz JE, Agus DB, MacCoss M, Tabb DL, Mallick P (2012). “A cross-platform toolkit for mass spectrometry and proteomics.” Nature Biotechnology, 30(10), 918--920. https://doi.org/10.1038/nbt.2377.

Examples

## The full file path
#> file = E:/ms_file.d

## Options to centroid MS1 and MS2 data in file
#> optList <- list(filter = "peakPicking vendor msLevel=1-2")

## Convert to mzML
#> convert_ms_files(file, outfile = "mzML", optList = optList)